+++ README.rst
@@ -136,12 +136,12 @@ or ``useomp=1`` to the end of the second line as shown above.
 Attribution
 ===========
 
-.. image:: https://img.shields.io/badge/JOSS-doi-brightgreen
-    :target: https://joss.theoj.org/papers/?/status.svg
+.. image:: https://joss.theoj.org/papers/a8325e3897dd726d9df42286bf72d19f/status.svg
+    :target: https://joss.theoj.org/papers/a8325e3897dd726d9df42286bf72d19f
     :alt: JOSS
 
-.. image:: https://img.shields.io/badge/arXiv-yymm.%3F-b31b1b
-    :target: https://arxiv.org/abs/?.?
+.. image:: https://img.shields.io/badge/arXiv-2304.03643-b31b1b
+    :target: https://arxiv.org/abs/2304.03643
     :alt: arXiv
 
 .. image:: https://img.shields.io/badge/doi-10.1093%2Fmnras%2Fsty3249-informational
