+++ .github/workflows/cd_arm64.yml
@@ -1,4 +1,4 @@
-name: continuous-delivery
+name: continuous-delivery-arm64
 
 on:
   push:
@@ -7,67 +7,41 @@ on:
       - 'v[0-9]+.[0-9]+.[0-9]+.post[0-9]+'
       - 'v[0-9]+.[0-9]+.[0-9]+rc[0-9]+'
     paths:
-      - deploy/pkg/conda_recipe/**
-      - .github/workflows/cd.yml
+      - deploy/pkg/conda_recipe_arm64/**
+      - .github/workflows/cd_arm64.yml
 
 concurrency:
   group: ${{ github.workflow }}-${{ github.ref }}
   cancel-in-progress: true
 
 jobs:
-  build_sdist:
-    name: Build sdist tarball
-
-    runs-on: ubuntu-latest
-
-    timeout-minutes: 20
-
-    steps:
-      - name: Checkout
-        uses: actions/checkout@v3
-
-      - name: Set up Python 3
-        uses: actions/setup-python@v4
-        with:
-          python-version: '3.x'
-
-      - name: Install packaging requirements
-        run: python -m pip install --upgrade build twine
-
-      - name: Build source distribution
-        run: python -m build --sdist --outdir dist/ .
-
-      - name: Verify source distribution
-        run: python -m twine check --strict dist/*
-
-      - name: Save tarball to sdist
-        uses: actions/upload-artifact@v3
-        with:
-          name: sdist_tar_${{ github.ref_name }}
-          path: dist/*.tar.gz
-
-      - name: Save tarball to pypi_dist
-        uses: actions/upload-artifact@v3
-        with:
-          name: pypi_dist_${{ github.ref_name }}
-          path: dist/*.tar.gz
-
-  build_bdist:
-    name: Build bdist wheel
+  build_bdist_arm64:
+    name: Build bdist wheel (arm64)
 
     strategy:
       fail-fast: false
       matrix:
         # Use oldest supported OS for backward compatibility.
         os: [ubuntu-20.04, macos-11]
+        # Skip non-arm64 macOS runner (as cross compilation is not working).
+        exclude:
+          - os: macos-11
 
     runs-on: ${{ matrix.os }}
 
-    timeout-minutes: 60
+    timeout-minutes: 180
 
     steps:
       - name: Checkout
         uses: actions/checkout@v3
+
+      # Use QEMU for non-arm64 Linux runner.
+      - name: Set up QEMU (Linux)
+        if: runner.os == 'Linux'
+        uses: docker/setup-qemu-action@v2
+        with:
+          # Specify arm64 architecture.
+          platforms: linux/arm64
 
       - name: Install packaging requirements
         run: python -m pip install --upgrade twine
@@ -76,36 +50,44 @@ jobs:
         uses: pypa/cibuildwheel@v2.12.1
         with:
           package-dir: .
-          output-dir: dist/
+          output-dir: dist_arm64/
           config-file: pyproject.toml
         env:
-          # Default to GCC compiler and OpenMP on macOS.
+          # Optional: Specify arm64 architecture (as non-native).
+          CIBW_ARCHS_LINUX: aarch64
+          CIBW_ARCHS_MACOS: arm64
+          # Default to LLVM compiler and OpenMP on macOS.
           CIBW_ENVIRONMENT_MACOS: >-
-            PY_CXX=$(brew --prefix gcc)/bin/g++-12
-            PY_CXXFLAGS=-I$(brew --prefix)/include
-            PY_LDFLAGS=-L$(brew --prefix)/lib
+            PY_CXX=$(brew --prefix llvm)/bin/clang++
+            PY_CXXFLAGS="-I$(brew --prefix)/include"
+            PY_LDFLAGS="-L$(brew --prefix)/lib"
[…]
