### 2023-04-12 8e30e44
Add conda package compression. Add manual workflow triggers
