### 2023-04-12 c8ea5e4
Add conda package compression. Add manual workflow triggers
