+++ deploy/pkg/conda_recipe/meta.yaml
@@ -1,5 +1,5 @@
 {% set name = 'Triumvirate' %}
-{% set version = environ.get('GIT_DESCRIBE_TAG') %}
+{% set version = environ.get('GIT_DESCRIBE_TAG', 'v0.1') %}
 
 package:
   name: "{{ name|lower }}"
