### 2023-04-12 717e851
Add commented syntax shortcut
