+++ deploy/pkg/conda_recipe_arm64/meta.yaml
@@ -7,6 +7,7 @@ package:
 
 source:
   git_url: ../../../
+  # git_rev: v0.1.2
   # path: ../../../
 
 build:
