### 2023-04-13 dd9da21
[gh-actions] Bump knicknic/os-specific-run from 1.0.3 to 1.0.4 (#5)
Bumps [knicknic/os-specific-run](https://github.com/knicknic/os-specific-run) from 1.0.3 to 1.0.4.
- [Release notes](https://github.com/knicknic/os-specific-run/releases)
- [Commits](https://github.com/knicknic/os-specific-run/compare/v1.0.3...v1.0.4)

---
updated-dependencies:
- dependency-name: knicknic/os-specific-run
  dependency-type: direct:production
  update-type: version-update:semver-patch
...

Signed-off-by: dependabot[bot] <support@github.com>
Co-authored-by: dependabot[bot] <49699333+dependabot[bot]@users.noreply.
[…]
