### 2023-04-13 3c5bcbd
Fix `make` recursive variable reference error
