+++ setup.py
@@ -1,11 +1,11 @@
 """Set up Triumvirate and its Cythonised extension modules.
 
 """
-import logging
 import os
 import platform
 import sys
 from copy import deepcopy
+from distutils.log import _global_log as distutils_logger
 from multiprocessing import cpu_count
 from setuptools import setup
 from setuptools.command.build_clib import build_clib
@@ -15,9 +15,6 @@ from Cython.Build import cythonize
 from Cython.Distutils import build_ext, Extension
 from extension_helpers._openmp_helpers import check_openmp_support
 from extension_helpers._setup_helpers import pkg_config
-
-
-distutils_logger = logging.getLogger()  # recycled from :mod:`distutils._log`
 
 
 # ========================================================================
@@ -45,7 +42,8 @@ def get_pkg_dir(topdir="src"):
     Parameters
     ----------
     topdir : str, optional
-        Top directory.
+        Top-level (sub-)directory containing the package at the
+        repository root.
 
     Returns
     -------
@@ -62,7 +60,7 @@ def get_pkg_include_dir(subdir="include"):
     Parameters
     ----------
     subdir : str, optional
-        Subdirectory of C++ headers in the package directory.
+        Subdirectory containing C++ headers in the package directory.
 
     Returns
     -------
@@ -73,13 +71,13 @@ def get_pkg_include_dir(subdir="include"):
     return os.path.join(get_pkg_dir(), subdir)
 
 
-def get_pkg_src_dir(subdir="src/modules"):
+def get_pkg_src_dir(subdir="src"):
     """Get package C++ source directory.
 
     Parameters
     ----------
     subdir : str, optional
-        Subdirectory of C++ sources in the package directory.
+        Subdirectory containing C++ sources in the package directory.
 
     Returns
     -------
@@ -98,9 +96,10 @@ class BuildExt(build_ext):
     """Modified :class:`Cython.Distutils.build_ext`.
 
     """
+
     def finalize_options(self):
         """Modify parallelisation option in
-        :meth:`Cython.Distutils.build_ext.build_ext.build_extensions`.
+        :meth:`Cython.Distutils.build_ext.finalize_options`.
 
         """
         super().finalize_options()
@@ -111,7 +110,7 @@ class BuildExt(build_ext):
 
     def build_extensions(self):
         """Modify compiler and compilation configuration in
-        :meth:`Cython.Distutils.build_ext.build_ext.build_extensions`.
+        :meth:`Cython.Distutils.build_ext.build_extensions`.
 
         """
         OPTS_TO_REMOVE = ['-Wstrict-prototypes',]  # noqa: E231
@@ -138,6 +137,7 @@ class BuildClib(build_clib):
     """Modified :class:`setuptools.command.build_clib.build_clib`.
 
     """
+
     def build_libraries(self, libraries):
         """Modify compiler and compilation configuration in
         :meth:`setuptools.command.build_clib.build_clib.build_libraries`.
@@ -159,7 +159,7 @@ class BuildClib(build_clib):
 
 
 def get_build_num_procs():
-    """Get build parallel processes.
+    """Get the number of build parallel processes.
 
     Returns
     -------
@@ -183,7 +183,7 @@ def get_build_num_procs():
 
 
 def prioprint(*args, **kwargs):
-    """`print` in high priority.
+    """`print` with high priority.
 
     """
     print(*args, file=sys.stderr, **kwargs)
@@ -195,9 +195,9 @@ def display_py_environs():
     """
[…]
