+++ Makefile
@@ -7,8 +7,9 @@
 # Configuration
 # ========================================================================
 
-PROGNAME = triumvirate
-LIBNAME = trv
+PROGNAME := triumvirate
+LIBNAME := trv
+
 
 # ------------------------------------------------------------------------
 # Directories
@@ -17,78 +18,97 @@ LIBNAME = trv
 # Repository root
 DIR_ROOT := $(shell pwd)
 
-# Package, build and test directories
+# Package, build, test and dist directories
 DIR_PKG := ${DIR_ROOT}/src/${PROGNAME}
 DIR_BUILD := ${DIR_ROOT}/build
 DIR_TESTS := ${DIR_ROOT}/tests
+DIR_DIST := ${DIR_ROOT}/dist
 
 # Package subdirectories
 DIR_PKG_INCLUDE := ${DIR_PKG}/include
 DIR_PKG_SRC := ${DIR_PKG}/src
+DIR_PKG_SRCPROG := ${DIR_PKG}/main
 
 # Build subdirectories
-DIR_BUILDLIB := ${DIR_BUILD}/lib
 DIR_BUILDOBJ := ${DIR_BUILD}/obj
+DIR_BUILDLIB := ${DIR_BUILD}/lib
 DIR_BUILDBIN := ${DIR_BUILD}/bin
 
 # Test subdirectories
 DIR_TESTBUILD := ${DIR_TESTS}/test_build
 DIR_TESTOUT := ${DIR_TESTS}/test_output
 
-# Package source module subdirectory
-DIR_PKG_SRCMODULES := ${DIR_PKG_SRC}/modules
-
 
 # ------------------------------------------------------------------------
-# Make Options
+# Options
 # ------------------------------------------------------------------------
 
-# Extract the '-j' or '--jobs' option (possibly empty). Note the space
-# added after ``${MAKEFLAGS}``.
-MAKEFLAGS_JOBS=$(shell echo "${MAKEFLAGS} " | grep -Eo "\-j[[:digit:][:space:]]*[^a-z[:punct:]]")
-
+# Extract the '-j' or '--jobs' option (possibly empty).
+# Note the space added after ``${MAKEFLAGS}``.
+PATTERN_JOBS = "\-j[[:digit:][:space:]]*[^a-z[:punct:]]"
+MAKEFLAGS_JOBS = $(shell echo "${MAKEFLAGS} " | grep -Eo ${PATTERN_JOBS})
 
 # ------------------------------------------------------------------------
-# OS-dependent Compilation
+# Compilation
 # ------------------------------------------------------------------------
+
+# -- OS ------------------------------------------------------------------
+
+OS := $(shell uname -s)
+
 
 # -- Compiler ------------------------------------------------------------
 
-# Assume GCC compiler by default. [adapt]
-ifeq ($(shell uname -s), Linux)
+# Assume explicitly GCC compiler by default. [adapt]
+ifeq (${OS}, Linux)
 
 CXX ?= g++
 
-else ifeq ($(shell uname -s), Darwin)
+else ifeq (${OS}, Darwin)
 
 # Use GCC compiler from Homebrew (brew formula 'gcc').
-# The compiler binary provided may have the suffix '-<version>';
-# check the version number with `brew info gcc` (here 'g++-11' is assumed).
-CXX_DIR = $(shell brew --prefix gcc)/bin
[…]
