+++ .github/workflows/cd.yml
@@ -33,6 +33,10 @@ on:
         type: boolean
         default: false
         required: false
+      version_tag:
+        description: 'Version tag for delivery'
+        type: string
+        required: false
 
 concurrency:
   group: ${{ github.workflow }}-${{ github.ref }}
@@ -160,6 +164,8 @@ jobs:
     steps:
       - name: Checkout
         uses: actions/checkout@v3
+        with:
+          fetch-depth: 0
 
       - name: Set up (Mini)conda
         uses: conda-incubator/setup-miniconda@v2
@@ -175,6 +181,22 @@ jobs:
           conda_root=$(conda config --show root_prefix | sed 's/root_prefix: //')
           conda_env=${conda_root}/envs/conda_bld
           export PATH=$PATH:${conda_root}/bin:${conda_env}/bin
+
+      - name: Override version
+        uses: knicknic/os-specific-run@v1.0.4
+        env:
+          recipe_file: deploy/pkg/conda_recipe/meta.yaml
+        with:
+          linux: |
+            vers_tag=${{ github.event.inputs.version_tag }}
+            if [[ ! -z ${vers_tag} ]]; then
+              sed -i "s/# git_rev: v.*/git_rev: ${vers_tag}/g" ${recipe_file}
+            fi
+          macos: |
+            vers_tag=${{ github.event.inputs.version_tag }}
+            if [[ ! -z ${vers_tag} ]]; then
+              sed -i '' "s/# git_rev: v.*/git_rev: ${vers_tag}/g" ${recipe_file}
+            fi
 
       - name: Build conda package
         env:
