+++ docs/dependencies.txt
@@ -6,3 +6,4 @@ sphinx-book-theme
 sphinx-copybutton
 sphinx-tabs
 sphinx-togglebutton
+setuptools-scm
