+++ README.rst
@@ -181,10 +181,18 @@ User feedback and contributions are very welcome. Please refer to the
 <https://github.com/MikeSWang/Triumvirate/blob/main/CONTRIBUTING.md>`_.
 
 
+Discussions
+===========
+
+A `community forum <https://github.com/MikeSWang/Triumvirate/discussions>`_
+for users and developers is hosted on GitHub, where you can receive
+announcements, post questions, share ideas and get updates.
+
+
 Releases
 ========
 
-Changes in current and past releases are listed in the `change log
+Release notes are included in the `change log
 <https://github.com/MikeSWang/Triumvirate/blob/main/CHANGELOG.md>`_.
 
 
