+++ deploy/pkg/autoinc_version.sh
@@ -1,6 +1,6 @@
 #!/usr/bin/env bash
 #
-# @file autoinc_vers.sh
+# @file autoinc_version.sh
 # @author Mike S Wang
 # @brief Increment fallback version number based on git tags.
 #
@@ -18,11 +18,11 @@ get_latest_release () {
     printf $(printf ${release_latest} | sed -e "s/^v//")
 }
 
-# @brief Set fallback version number in relevant files.
+# @brief Set fallback version number in ``__init__.py`` file.
 #
 # @arg Version number.
 #
-set_fallback_version () {
+set_fallback_version_initpy () {
     vers=$1
     versfile=src/triumvirate/__init__.py
     versline="__version__ = '.*'.*"
@@ -30,5 +30,19 @@ set_fallback_version () {
     sed -i "s/${versline}/${versfill}/g" ${versfile}
 }
 
+# @brief Set fallback version number in ``meta.yaml`` files.
+#
+# @arg Version number.
+#
+set_fallback_version_metayaml () {
+    vers=$1
+    versline="{% set version = environ.get('GIT_DESCRIBE_TAG', .*) %}"
+    versfill="{% set version = environ.get('GIT_DESCRIBE_TAG', 'v${vers}') %}"
+    for versfile in $(find deploy/pkg -type f -name "meta.yaml"); do
+        sed -i "s/${versline}/${versfill}/g" ${versfile}
+    done
+}
+
 # Increment fallback version based on the latest-release git tag.
-set_fallback_version $(get_latest_release)
+set_fallback_version_initpy $(get_latest_release)
+set_fallback_version_metayaml $(get_latest_release)
