### 2023-04-20 9c79ca1
Add dir path to error message
