### 2023-04-21 5eda64f
Add (un)installation options
