### 2023-05-25 f676754
Add `pypower`-like normalisation for 2-pt stats
