+++ application/extern/pypower/powspec_pyp.py
@@ -1,0 +1,269 @@
+"""Measure power spectrum using `pypower`.
+
+"""
+import os.path as osp
+from argparse import ArgumentParser
+from itertools import repeat
+from pathlib import Path
+from pprint import pformat
+from tempfile import TemporaryDirectory
+
+import numpy as np
+from pypower import CatalogFFTPower, normalization_from_nbar
+
+
+def configure():
+    """Parse command-line arguments to runtime parameters.
+
+    Returns
+    -------
+    :class:`argparse.Namespace`
+        Parameter namespace.
+
+    """
+    parser = ArgumentParser(
+        prog="powspec_pyp",
+        description="Measure power spectrum using `pypower`."
+    )
+
+    parser.add_argument(
+        '--input-dir', type=str, default="./",
+        help="input (catalogue) directory"
+    )
+    parser.add_argument(
+        '--output-dir', type=str, default="./",
+        help="output (measurement) directory"
+    )
+    parser.add_argument(
+        '--data-ctlgfile', type=str, required=True,
+        help="data-source catalogue file"
+    )
+    parser.add_argument(
+        '--rand-ctlgfile', type=str, default=None,
+        help="random-source catalogue file"
+    )
+    parser.add_argument(
+        '--ctlg-fields', type=str, nargs='+', required=True,
+        help="catalogue fields"
+    )
+    parser.add_argument(
+        '--output-file', type=str, required=True,
+        help="output measurement file name"
+    )
+
+    parser.add_argument(
+        '--boxsize', type=float, nargs='+', required=True,
+        help="box size for all dimensions"
+    )
+    parser.add_argument(
+        '--ngrid', type=int, nargs='+', required=True,
+        help="mesh grid number"
+    )
+    parser.add_argument(
+        '--scheme', type=str, default='pcs',
+        help="mesh assignment scheme"
+    )
+    parser.add_argument(
+        '--interlace', action='store_true',
+        help="interlacing switch"
+    )
+
+    parser.add_argument(
+        '--multipole', type=int, required=True,
+        help="power spectrum multipole degree"
+    )
+    parser.add_argument(
+        '--range', type=float, nargs=2, required=True,
+        help="wavenumber range (in h/Mpc)"
+    )
+    parser.add_argument(
+        '--num-bins', type=int, required=True,
+        help="number of bins"
+    )
+
+    parser.add_argument(
+        '--norm-only', action='store_true',
+        help="normalisation-only switch"
+    )
+    parser.add_argument(
+        '--no-norm', action='store_true',
+        help="no-normalisation switch"
+    )
+
+    config = parser.parse_args()
+
+    if len(config.boxsize) == 1:
+        config.boxsize = [config.boxsize] * 3
+    if len(config.ngrid) == 1:
+        config.ngrid = [config.ngrid] * 3
+
+    return config
+
+
+def run_pypower(config):
+    """Run `pypower` to measure the power spectrum.
+
+    Parameters
+    ----------
+    config : :class:`argparse.Namespace`
+        Parameters.
+
+    Returns
+    -------
+    :class:`pypower.CatalogFFTPower`
+        Power spectrum measurements.
+
+    """
+    def get_dtype():
+        return list(zip(config.ctlg_fields, repeat(float)))
+
+    # Load catalogues.
+    data_ctlgpath = osp.join(config.input_dir, config.data_ctlgfile)
+    if config.rand_ctlgfile is None:
+        rand_ctlgpath = None
+    else:
[…]

[…]
