+++ .github/workflows/cd.yml
@@ -109,7 +109,7 @@ jobs:
         run: python -m pip install --upgrade twine
 
       - name: Build built distribution
-        uses: pypa/cibuildwheel@v2.12.3
+        uses: pypa/cibuildwheel@v2.13.0
         with:
           package-dir: .
           output-dir: dist/
