### 2023-05-31 c73e623
Upgrade API for added mix-mesh normalisation
Match Python and C++ names. Add default and copy con-/de-structor. Update class attributes. Update logging and headers. Fix typos
