### 2023-05-31 708e9c5
Enforce normalisation mesh cell size. Update output header
