### 2023-05-31 ad1121a
Add variable to ``checkopts``
