### 2023-05-31 c7bb277
Make syntax non-bash-specific
