+++ .github/workflows/cd.yml
@@ -117,7 +117,7 @@ jobs:
         env:
           # Default to GCC compiler and OpenMP on macOS.
           CIBW_ENVIRONMENT_MACOS: >-
-            PY_CXX=$(brew --prefix gcc)/bin/g++-12
+            PY_CXX=$(find $(brew --prefix gcc)/bin -type f -name 'g++*')
             PY_CXXFLAGS=-I$(brew --prefix)/include
             PY_LDFLAGS=-L$(brew --prefix)/lib
 
