### 2023-05-31 749bf05
Fix derived parameter updating
