### 2023-06-01 df00160
Update syntax. Wrap new C++ normalisation function in Cython
