+++ src/triumvirate/_twopt.pyx
@@ -7,6 +7,7 @@ Declare and interface with two-point correlator algorithms.
 """
 from cython.operator cimport dereference as deref
 from libc.stdlib cimport free, malloc
+from libcpp.string cimport string
 
 import numpy as np
 cimport numpy as np
@@ -25,18 +26,37 @@ cdef extern from "include/twopt.hpp":
     # Normalisation
     # --------------------------------------------------------------------
 
+    double calc_powspec_normalisation_from_particles_cpp \
+        "trv::calc_powspec_normalisation_from_particles" (
+            CppParticleCatalogue& particles,
+            double alpha
+        ) except +
+
     double calc_powspec_normalisation_from_mesh_cpp \
         "trv::calc_powspec_normalisation_from_mesh" (
-            CppParticleCatalogue& catalogue,
+            CppParticleCatalogue& particles,
             CppParameterSet& params,
             double alpha
         )
 
-    double calc_powspec_normalisation_from_particles_cpp \
-        "trv::calc_powspec_normalisation_from_particles" (
-            CppParticleCatalogue& catalogue,
+    double calc_powspec_normalisation_from_meshes_cpp \
+        "trv::calc_powspec_normalisation_from_meshes" (
+            CppParticleCatalogue& particles_data,
+            CppParticleCatalogue& particles_rand,
+            CppParameterSet& params,
             double alpha
-        ) except +
+        )
+
+    double calc_powspec_normalisation_from_meshes_cpp \
+        "trv::calc_powspec_normalisation_from_meshes" (
+            CppParticleCatalogue& particles_data,
+            CppParticleCatalogue& particles_rand,
+            CppParameterSet& params,
+            double alpha,
+            double padding,
+            double cellsize,
+            string assignment
+        )
 
 
     # --------------------------------------------------------------------
@@ -44,8 +64,8 @@ cdef extern from "include/twopt.hpp":
     # --------------------------------------------------------------------
 
     PowspecMeasurements compute_powspec_cpp "trv::compute_powspec" (
-        CppParticleCatalogue& particles_data,
-        CppParticleCatalogue& particles_rand,
+        CppParticleCatalogue& catalogue_data,
+        CppParticleCatalogue& catalogue_rand,
         LineOfSight* los_data,
         LineOfSight* los_rand,
         CppParameterSet& params,
@@ -54,8 +74,8 @@ cdef extern from "include/twopt.hpp":
     )
 
     TwoPCFMeasurements compute_corrfunc_cpp "trv::compute_corrfunc" (
-        CppParticleCatalogue& particles_data,
-        CppParticleCatalogue& particles_rand,
+        CppParticleCatalogue& catalogue_data,
+        CppParticleCatalogue& catalogue_rand,
         LineOfSight* los_data,
         LineOfSight* los_rand,
         CppParameterSet& params,
@@ -65,7 +85,7 @@ cdef extern from "include/twopt.hpp":
 
     PowspecMeasurements compute_powspec_in_gpp_box_cpp \
         "trv::compute_powspec_in_gpp_box" (
-            CppParticleCatalogue& particles_data,
+            CppParticleCatalogue& catalogue_data,
             CppParameterSet& params,
             CppBinning& kbinning,
             double norm_factor
@@ -73,7 +93,7 @@ cdef extern from "include/twopt.hpp":
 
     TwoPCFMeasurements compute_corrfunc_in_gpp_box_cpp \
         "trv::compute_corrfunc_in_gpp_box" (
-            CppParticleCatalogue& particles_data,
+            CppParticleCatalogue& catalogue_data,
             CppParameterSet& params,
[…]
