+++ src/triumvirate/_threept.pyx
@@ -28,13 +28,13 @@ cdef extern from "include/threept.hpp":
 
     double calc_bispec_normalisation_from_particles_cpp \
         "trv::calc_bispec_normalisation_from_particles" (
-            CppParticleCatalogue& catalogue,
+            CppParticleCatalogue& particles,
             double alpha
         ) except +
 
     double calc_bispec_normalisation_from_mesh_cpp \
         "trv::calc_bispec_normalisation_from_mesh" (
-            CppParticleCatalogue& catalogue,
+            CppParticleCatalogue& particles,
             CppParameterSet& params,
             double alpha
         )
@@ -45,8 +45,8 @@ cdef extern from "include/threept.hpp":
     # --------------------------------------------------------------------
 
     BispecMeasurements compute_bispec_cpp "trv::compute_bispec" (
-        CppParticleCatalogue& particles_data,
-        CppParticleCatalogue& particles_rand,
+        CppParticleCatalogue& catalogue_data,
+        CppParticleCatalogue& catalogue_rand,
         LineOfSight* los_data,
         LineOfSight* los_rand,
         CppParameterSet& params,
@@ -55,8 +55,8 @@ cdef extern from "include/threept.hpp":
     )
 
     ThreePCFMeasurements compute_3pcf_cpp "trv::compute_3pcf" (
-        CppParticleCatalogue& particles_data,
-        CppParticleCatalogue& particles_rand,
+        CppParticleCatalogue& catalogue_data,
+        CppParticleCatalogue& catalogue_rand,
         LineOfSight* los_data,
         LineOfSight* los_rand,
         CppParameterSet& params,
@@ -66,7 +66,7 @@ cdef extern from "include/threept.hpp":
 
     BispecMeasurements compute_bispec_in_gpp_box_cpp \
         "trv::compute_bispec_in_gpp_box" (
-            CppParticleCatalogue& particles_data,
+            CppParticleCatalogue& catalogue_data,
             CppParameterSet& params,
             CppBinning& kbinning,
             double norm_factor
@@ -74,7 +74,7 @@ cdef extern from "include/threept.hpp":
 
     ThreePCFMeasurements compute_3pcf_in_gpp_box_cpp \
         "trv::compute_3pcf_in_gpp_box" (
-            CppParticleCatalogue& particles_data,
+            CppParticleCatalogue& catalogue_data,
             CppParameterSet& params,
             CppBinning& rbinning,
             double norm_factor
@@ -82,7 +82,7 @@ cdef extern from "include/threept.hpp":
 
     ThreePCFWindowMeasurements compute_3pcf_window_cpp \
         "trv::compute_3pcf_window" (
-            CppParticleCatalogue& particles_rand,
+            CppParticleCatalogue& catalogue_rand,
             LineOfSight* los_rand,
             CppParameterSet& params,
             CppBinning& rbinning,
@@ -93,8 +93,8 @@ cdef extern from "include/threept.hpp":
 
     # BispecMeasurements compute_bispec_for_los_choice_cpp \
     #     "trv::compute_bispec_for_los_choice" (
-    #         CppParticleCatalogue& particles_data,
-    #         CppParticleCatalogue& particles_rand,
+    #         CppParticleCatalogue& catalogue_data,
+    #         CppParticleCatalogue& catalogue_rand,
     #         LineOfSight* los_data,
     #         LineOfSight* los_rand,
     #         int los_choice,
@@ -104,27 +104,27 @@ cdef extern from "include/threept.hpp":
     #     )
 
 
+def _calc_bispec_normalisation_from_particles(
[…]
