### 2023-06-01 b7de60f
Add normalisation options and modify relevant I/O
