+++ src/triumvirate/twopt.py
@@ -32,6 +32,7 @@ import numpy as np
 
 from triumvirate._twopt import (
     _calc_powspec_normalisation_from_mesh,
+    _calc_powspec_normalisation_from_meshes,
     _calc_powspec_normalisation_from_particles,
     _compute_corrfunc,
     _compute_corrfunc_in_gpp_box,
@@ -46,6 +47,12 @@ from triumvirate.parameters import (
     fetch_paramset_template,
     ParameterSet,
 )
+
+
+# Use default parameters for mixed-mesh normalisation in `pypower`.
+PADDING = .10
+CELLSIZE = 10.
+ASSIGNMENT = 'cic'
 
 
 def _amalgamate_parameters(paramset=None, params_sampling=None,
@@ -180,7 +187,8 @@ def _get_measurement_filename(paramset):
     )
 
 
-def _print_measurement_header(paramset, norm_factor, norm_factor_alt):
+def _print_measurement_header(paramset, norm_factor_part, norm_factor_mesh,
+                              norm_factor_meshes):
     """Print two-point statistic measurement header including
     sampling parameters, normalisation factors and data table columns.
 
@@ -188,8 +196,8 @@ def _print_measurement_header(paramset, norm_factor, norm_factor_alt):
     ----------
     paramset : :class:`~triumvirate.parameters.ParameterSet`
         Parameter set.
-    norm_factor, norm_factor_alt : double
-        Normalisation factor used and the alternative.
+    norm_factor_part, norm_factor_mesh, norm_factor_meshes : float
+        Normalisation factors.
 
     Returns
     -------
@@ -203,12 +211,14 @@ def _print_measurement_header(paramset, norm_factor, norm_factor_alt):
         two-point statistics.
 
     """
+    if paramset['norm_convention'] == 'none':
+        norm_factor = 1.
     if paramset['norm_convention'] == 'particle':
-        norm_convention = 'particle'
-        norm_convention_alt = 'mesh'
+        norm_factor = norm_factor_part
     if paramset['norm_convention'] == 'mesh':
-        norm_convention = 'mesh'
-        norm_convention_alt = 'particle'
+        norm_factor = norm_factor_mesh
+    if paramset['norm_convention'] == 'meshes':
+        norm_factor = norm_factor_meshes
 
     if paramset['npoint'] != '2pt':
         raise ValueError(
@@ -241,10 +251,12 @@ def _print_measurement_header(paramset, norm_factor, norm_factor_alt):
         "Mesh assignment and interlacing: {}, {}".format(
             paramset['assignment'], paramset['interlace']
         ),
-        "Normalisation factor: "
-        "{:.9e} ({}-based, used), {:.9e} ({}-based, alternative)".format(
-            norm_factor, norm_convention,
-            norm_factor_alt, norm_convention_alt
+        "Normalisation factor: {:.9e} ({})".format(
+            norm_factor, paramset['norm_convention']
+        ),
+        "Normalisation factor alternatives: "
+        "{:.9e} (particle), {:.9e} (mesh), {:.9e} (mesh-mixed)".format(
+            norm_factor_part, norm_factor_mesh, norm_factor_meshes
         ),
         ", ".join([
             "[{:d}] {}".format(colidx, colname)
@@ -463,25 +475,46 @@ def _compute_2pt_stats_survey_like(twopt_algofunc,
     if logger:
         logger.info("Alpha contrast: %.6e.", alpha)
 
-    if paramset['norm_convention'] == 'particle':
-        norm_factor = _calc_powspec_normalisation_from_particles(
-            particles_rand, alpha
+    norm_factor_part = _calc_powspec_normalisation_from_particles(
+        particles_rand, alpha
+    )
[…]
