### 2023-06-01 7c87294
Rerun notebooks. Update syntax and fix typos
