### 2023-06-02 f5f0352
Fix `npair` counting. Fix indentation
