+++ deploy/pkg/conda_recipe_arm64/meta.yaml
@@ -1,5 +1,5 @@
 {% set name = 'Triumvirate' %}
-{% set version = environ.get('GIT_DESCRIBE_TAG', 'v0.1') %}
+{% set version = environ.get('GIT_DESCRIBE_TAG', 'v0.2.0') %}
 
 package:
   name: "{{ name|lower }}"
