### 2023-06-16 9f62a30
Remove reality-condition factor for generality
