+++ .github/workflows/ci.yml
@@ -36,16 +36,6 @@ on:
       - .github/workflows/ci.yml
   pull_request_review:
     types: [submitted, edited]
-    paths:
-      - src/**
-      - tests/**
-      - setup.py
-      - setup.cfg
-      - pyproject.toml
-      - MANIFEST.in
-      - Makefile
-      - .gitignore
-      - .github/workflows/ci.yml
   workflow_dispatch:
 
 concurrency:
