+++ .github/ISSUE_TEMPLATE/bug-report.md
@@ -40,8 +40,8 @@ to help explain your problem.
 
 - Package version (e.g. 1.0.0)
 - OS (e.g. Linux, macOS)
-- Python version (e.g. 3.8) and environment (e.g. conda)
-- C++ compiler version (e.g. GCC g++, LLVM clang++) and dependency
+- Python version (e.g. 3.8) and environment (e.g. Conda)
+- C++ compiler version (e.g. GCC ``g++``, LLVM Clang++) and dependency
   libraries (e.g. ``libgomp``, ``libomp``)
 
 
