### 2023-06-21 ec82b1a
Add explicit OMP switches to `setup.py`
