+++ deploy/pkg/conda_recipe_arm64/meta.yaml
@@ -18,6 +18,11 @@ build:
     # Default to LLVM compiler and OpenMP on macOS.
     - export PY_CXXFLAGS_OMP="-Xpreprocessor -fopenmp"  # [osx]
     - export PY_LDFLAGS_OMP="-lomp"                     # [osx]
+    - echo $(which {{ PYTHON }})
+    - echo $(which python)
+    - echo $(conda list | grep python)
+    - echo $({{ PYTHON }} --version)
+    - echo $(python --version)
     - {{ PYTHON }} -m pip install . -vvv
   script_env:
     - PY_BUILD_PARALLEL="-j"
@@ -26,8 +31,8 @@ requirements:
   build:
     - {{ compiler('cxx') }}
     - llvm-openmp                         # [osx]
-    - python                              # [build_platform != target_platform]
     - cross-python_{{ target_platform }}  # [build_platform != target_platform]
+    - python                              # [build_platform != target_platform]
     - cython                              # [build_platform != target_platform]
     - numpy                               # [build_platform != target_platform]
   host:
