+++ .github/workflows/cd_xp.yml
@@ -1,11 +1,11 @@
-# Provide continuous delivery (CD) for ARM64-based target platforms
-# from (non-native) build platforms.
+# Provide continuous delivery (CD) for target platforms
+# from non-native build platforms.
 #
-# CD deliverables currently include:
+# CD deliverables for target platforms (ARM64-based):
 # - (non-natively) built Python wheel distributions (for PyPI);
 # - (non-natively) built Conda packages (for Anaconda).
 #
-name: continuous-delivery-arm64
+name: continuous-delivery-xplatform
 
 on:
   push:
@@ -14,17 +14,17 @@ on:
       - 'v[0-9]+.[0-9]+.[0-9]+.post[0-9]+'
       - 'v[0-9]+.[0-9]+.[0-9]+rc[0-9]+'
     paths:
-      - deploy/pkg/conda_recipe_arm64/**
-      - .github/workflows/cd_arm64.yml
+      - deploy/pkg/conda_recipe_xp/**
+      - .github/workflows/cd_xp.yml
   workflow_dispatch:
     inputs:
-      run_build_bdist_arm64:
-        description: 'Run job build_bdist_arm64'
+      run_build_bdist_xplat:
+        description: 'Run job build_bdist_xplat'
         type: boolean
         default: false
         required: false
-      run_build_conda_arm64:
-        description: 'Run job build_conda_arm64'
+      run_build_conda_xplat:
+        description: 'Run job build_conda_xplat'
         type: boolean
         default: false
         required: false
@@ -38,8 +38,8 @@ concurrency:
   cancel-in-progress: true
 
 jobs:
-  build_bdist_arm64:
-    name: Build bdist wheel (ARM64)
+  build_bdist_xplat:
+    name: Build bdist wheel
 
     strategy:
       fail-fast: false
@@ -49,37 +49,39 @@ jobs:
 
     runs-on: ${{ matrix.os }}
 
-    timeout-minutes: 180
+    timeout-minutes: 60
 
+    # Allow manual trigger from inputs.
     if: >
       github.event_name != 'workflow_dispatch' ||
-      github.event.inputs.run_build_bdist_arm64 == 'true'
+      github.event.inputs.run_build_bdist_xplat == 'true'
 
     steps:
       - name: Checkout
         uses: actions/checkout@v3
 
-      # Use QEMU for non-ARM64 Linux runner.
+      # Use QEMU for non-native Linux runner.
       - name: Set up QEMU (Linux)
         if: runner.os == 'Linux'
         uses: docker/setup-qemu-action@v2
         with:
-          # Specify ARM64 architecture.
+          # Specify target architecture (ARM64).
           platforms: linux/arm64
 
-      # Adapt Homebrew for non-ARM64 macOS runner.
+      # Adapt Homebrew for non-native macOS runner.
       - name: Set up Homebrew (macOS)
         if: runner.os == 'macOS'
         run: |
-          # Install dependency formulae to match runner platform.
+          # (Re-)install (OpenMP) dependency formulae
+          # to match target platform (ARM64-based).
+          TARG_PLAT=arm64_big_sur
           for formula in gsl fftw; do
-            brew fetch --force --bottle-tag=arm64_big_sur $formula
-            brew install $(brew --cache --bottle-tag=arm64_big_sur $formula)
+            brew fetch --force --bottle-tag=${TARG_PLAT} $formula
+            brew install $(brew --cache --bottle-tag=${TARG_PLAT} $formula)
           done
-          # Reinstall OpenMP-dependency formulae to match runner platform.
           for formula in gcc libomp; do
-            brew fetch --force --bottle-tag=arm64_big_sur $formula
[…]
