+++ setup.cfg
@@ -37,7 +37,7 @@ setup_requires =
     Cython>=0.25
     oldest-supported-numpy
 install_requires =
-    numpy
+    numpy>=1.17
     scipy
     astropy>=4.2
     pyyaml>4
