### 2023-06-23 b982e93
Update fallback version and fix typos. Add cross-platform OpenMP for Linux conda build
