+++ .github/workflows/cd_xp.yml
@@ -49,7 +49,7 @@ jobs:
 
     runs-on: ${{ matrix.os }}
 
-    timeout-minutes: 60
+    timeout-minutes: 120
 
     # Allow manual trigger from inputs.
     if: >
