### 2023-06-23 5ff5551
Update OpenMP instructions
