+++ deploy/pkg/conda_recipe/meta.yaml
@@ -15,7 +15,9 @@ build:
   script:
     # Use Conda-provided compiler.
     - export PY_CXX=$CXX
-    # Default to LLVM compiler and OpenMP on macOS.
+    # Optional: Enforce OpenMP support.
+    - export PY_OMP=1
+    # Default to LLVM OpenMP on macOS.
     - export PY_CXXFLAGS_OMP="-Xpreprocessor -fopenmp"  # [osx]
     - export PY_LDFLAGS_OMP="-lomp"                     # [osx]
     - {{ PYTHON }} -m pip install . -vvv
