+++ deploy/pkg/conda_recipe_xp/meta.yaml
@@ -26,13 +26,13 @@ build:
 
 requirements:
   build:
+    - python                              # [build_platform != target_platform]
+    - cross-python_{{ target_platform }}  # [build_platform != target_platform]
+    - cython                              # [build_platform != target_platform]
+    - numpy >=1.17                        # [build_platform != target_platform]
     - {{ compiler('cxx') }}
     - libgomp                             # [linux]
     - llvm-openmp                         # [osx]
-    - cross-python_{{ target_platform }}  # [build_platform != target_platform]
-    - python                              # [build_platform != target_platform]
-    - cython                              # [build_platform != target_platform]
-    - numpy >=1.17                        # [build_platform != target_platform]
   host:
     - python
     - pip >=19.1
@@ -47,8 +47,6 @@ requirements:
     - extension-helpers
     - numpy >=1.17
   run:
-    - libgomp                        # [linux]
-    - llvm-openmp                    # [osx]
     - python >=3.8
     - {{ pin_compatible("numpy") }}
     - scipy
