### 2023-06-23 119e94a
Simplify conda-build recipes
