+++ deploy/pkg/conda_recipe_xp/meta.yaml
@@ -28,11 +28,8 @@ requirements:
   build:
     - python                              # [build_platform != target_platform]
     - cross-python_{{ target_platform }}  # [build_platform != target_platform]
-    - cython                              # [build_platform != target_platform]
     - numpy >=1.17                        # [build_platform != target_platform]
     - {{ compiler('cxx') }}
-    - libgomp                             # [linux]
-    - llvm-openmp                         # [osx]
   host:
     - python
     - pip >=19.1
