+++ deploy/pkg/conda_recipe/meta.yaml
@@ -27,8 +27,6 @@ build:
 requirements:
   build:
     - {{ compiler('cxx') }}
-    - libgomp                # [linux]
-    - llvm-openmp            # [osx]
   host:
     - python
     - pip >=19.1
