+++ deploy/pkg/conda_recipe_xp/conda_build_config.yaml
@@ -1,5 +1,5 @@
 target_platform:
   # - linux-64     # [linux]
   - linux-aarch64  # [linux]
-  # - osx-64       # [macos]
-  - osx-arm64      # [macos]
+  # - osx-64       # [osx]
+  - osx-arm64      # [osx]
