### 2023-06-29 21f9f30
Add conda variant. Maintain texts
