+++ deploy/pkg/autorel_anaconda.sh
@@ -40,7 +40,7 @@ if [[ ! -d ${DIST_DIR} ]]; then mkdir -p ${DIST_DIR}; fi
 conda build purge
 conda build --strict-verify --no-anaconda-upload ${RECIPE_DIR} \
   --output-folder ${DIST_DIR} \
-  --variants "{'python': ['3.8', '3.9', '3.10']}"
+  --variants "{'python': ['3.8', '3.9', '3.10', '3.11']}"
 
 # Transmute compression formats.
 find ${DIST_DIR} -name "*.tar.bz2" \
