+++ NOTES.md
@@ -2,7 +2,8 @@
 
 - The following code tags are in use:
     - 'RFE';
-    - 'CAVEAT', 'SEE';
+    - 'CAVEAT';
+    - 'SEE';
     - 'STYLE'.
 
 - The following code tags are available for use:
