+++ .github/PULL_REQUEST_TEMPLATE.md
@@ -34,6 +34,6 @@ Please describe the implementation details related to the issue(s).
 
 ## Checklist
 
-- [ ] Read the [contribution guidelines](../CONTRIBUTING.md)
-- [ ] Write unit tests
-- [ ] Write doc-strings and comments
+- [ ] Read the [contribution guidelines](../CONTRIBUTING.md).
+- [ ] Write unit tests.
+- [ ] Write doc-strings and comments.
