+++ Makefile
@@ -23,7 +23,7 @@ SCM_LOC_SCHEME ?= node-and-date
 # Repository root
 DIR_ROOT := $(shell pwd)
 
-# Package, build, test and dist directories
+# Package, build, test and distribution directories
 DIR_PKG := ${DIR_ROOT}/src/${PKGNAME}
 DIR_BUILD := ${DIR_ROOT}/build
 DIR_TESTS := ${DIR_ROOT}/tests
@@ -86,7 +86,7 @@ CXX ?= g++
 
 endif  # OS
 
-# Assume default achiver. [adapt]
+# Assume default archiver. [adapt]
 AR ?= ar
 ARFLAGS ?= -rcsv
 
@@ -265,11 +265,14 @@ PROGLIB := ${DIR_BUILDLIB}/lib${LIBNAME}.a
 
 install: cppinstall pyinstall
 
-cppinstall: cpplibinstall cppappbuild
+cppinstall: cppinstall_ cpplibinstall cppappbuild
 
-cpplibinstall: ${PROGLIB}
+cppinstall_:
+	@echo "Installing Triumvirate C++ library/program..."
 
-cppappbuild: ${PROGEXE}
+cpplibinstall: library
+
+cppappbuild: executable
 
 pyinstall:
 	@echo "Installing Triumvirate Python package ${WOMP} OpenMP (in pip dev mode)..."
@@ -279,7 +282,8 @@ uninstall: cppuninstall pyuninstall
 
 cppuninstall:
 	@echo "Uninstalling Triumvirate C++ library/program..."
-	find ${DIR_BUILD} -mindepth 1 -maxdepth 1 ! -name ".gitignore" -exec rm -r {} +
+	@echo "  ... removing builds..."
+	@find ${DIR_BUILD} -mindepth 1 -maxdepth 1 ! -name ".gitignore" -exec rm -r {} +
 
 pyuninstall:
 	@echo "Uninstalling Triumvirate Python package (in pip mode)..."
@@ -288,22 +292,39 @@ pyuninstall:
 
 # -- Components ----------------------------------------------------------
 
-${PROGEXE}: ${PROGOBJ} $(OBJS)
+.PHONY: executable library OBJS_
+
+executable: OBJS_ ${PROGEXE}
+
+${PROGEXE}: $(OBJS) ${PROGOBJ}
 	@echo "Compiling Triumvirate C++ program ${WOMP} OpenMP..."
-	if [ ! -d ${DIR_BUILDBIN} ]; then mkdir -p ${DIR_BUILDBIN}; fi
+	@if [ ! -d ${DIR_BUILDBIN} ]; then \
+	    echo "  making bin subdirectory in build directory..."; \
+	    mkdir -p ${DIR_BUILDBIN}; \
+	fi
 	$(CXX) $(CXXFLAGS) $^ -o $@ $(LDFLAGS) $(LDLIBS)
+
+library: OBJS_ ${PROGLIB}
 
 ${PROGLIB}: $(OBJS)
 	@echo "Creating Triumvirate C++ library ${WOMP} OpenMP..."
-	if [ ! -d ${DIR_BUILDLIB} ]; then mkdir -p ${DIR_BUILDLIB}; fi
+	@if [ ! -d ${DIR_BUILDLIB} ]; then \
+	    echo "  making lib subdirectory in build directory..."; \
+	    mkdir -p ${DIR_BUILDLIB}; \
+	fi
 	$(AR) $(ARFLAGS) $@ $^
 
+OBJS_:
[…]
