### 2023-07-04 2ebb3e6
Update cross-compilation workflow triggers
