### 2023-07-06 4fc92cf
Speed up spherical Bessel interpolator
