### 2023-07-25 MikeSWang/Triumvirate#22 @MikeSWang
[FEAT] Full 2D three-point statistics
## Is the requested feature related to an issue?

No.

## Summary

Allow the user to compute the full 2D three-point statistics without fixing the first coordinate.

## Alternatives

N/A

## Implementation

This should be straightforward but there are two points of consideration:
- the output measurement data array needs to be arranged to a specific layout, e.g. fix $k_1$ to the first bin and vary $k_2$ before moving to the next bin for $k_1$;
- the computation time for a single execution is likely to be $N_\mathrm{bin}$ times longer.

## Additional context

This is in response to the JOSS review suggestion (https://github.com/openjournals/joss-reviews/issues/5571#issuecomment-1638129119).
