### 2023-07-25 #5571 @MikeSWang
Thanks @wcoulton for your valuable feedback!

I just came back from a conference and will be travelling again soon, but I will incorporate your suggestions especially with regard to number 2 (you're right about it; I have just opened a new feature request https://github.com/MikeSWang/Triumvirate/issues/22).

A quick response to the last comment: the `examples` folder is a placeholder, and I meant to collect users' scripts for different use cases like a 'gallery', which may not be formal enough to be included in the official documentation :)

> Dear @MikeSWang,
>
> In general I think that this package is amazing! It has lots of useful functionality and it is easy to install and use. I can certainly see myself using this in future projects!
>
> I had four minor comments:
>
> 1. I think it would be useful to add a little more description to the returns of the functions in the documentation. A few function outputs are just described as "results – Measurement results". The actual returned dictionary has a few parts and it could be useful to either describe that there or link to the examples where the output is explained in detail.
> 2. Unless I have misunderstood, it seems necessary to repeatedly run the program, iteratively updating idx_bin, to compute all the elements in the full 2D bispectrum. For my applications, and I suspect that of some other users, the full bispectrum would often be of interest. Perhaps a function could be added that returns this?
> 3. Perhaps in the background material you could add a sentence or two more summarizing the window functions?
> 4. In your supporting paper, I think it may be useful to add a reference to, and discussion of, some other codes that have some similar functionality. The first one that jumps first to mind is:
>    https://github.com/oliverphilcox/Spectra-Without-Windows
>
> Super minor: there seems to be an example folder in the repo but it is empty?
>
> This package is wonderful, congratulations! My apologies for the slow review.
