### 2023-08-03 #5571 @alfonso-veropalumbo
Dear @MikeSWang

First of all, thank you for your patience. I'm sorry for this delayed answer.

I've had the chance to test Triumvirate, and I've particularly appreciated the ease of use, the completeness of the documentation and the high performance in dealing with this task.
Moreover, the advantage of Triumvirate is that it allows computation of the anisotropic three-point correlation function (3PCF), fully exploiting the information encoded in the higher orders moment of the galaxy density field. This aspect constitutes a new feature (apart from Triumvirate predecessor, Hitomi) and a critical advancement of the clustering field.

I have minor suggestions related to the paper:

1) quantify performances in the paper. 3PCF computation is generally thought to be (and for large samples, it is) an expensive procedure, and it could be helpful to report the performances of this specific task.
2) improve the comparison with other codes on the market (e.g. https://arxiv.org/abs/2105.08722). On this point, I have a doubt about the description of nbodykit made in the draft.
Looking at the documentation (https://nbodykit.readthedocs.io/en/latest/api/_autosummary/nbodykit.algorithms.threeptcf.html#nbodykit.algorithms.threeptcf.SimulationBox3PCF) nbodykit implements the Slepian 2015 procedure, which is not assigning spherical harmonics to the mesh grid. This procedure computes the spherical harmonics on a pair-to-pair basis. As a result, it should be more accurate but significantly slower.

Apart from these minor comments, thank you again for this work, which is important for the clustering community.
I hope packages like this will help people approach this field, which will gain more and more importance for clustering analyses in cosmology.

Sorry again for the delayed comments. I'm available in case further clarifications are needed.
