+++ .github/workflows/cd.yml
@@ -110,7 +110,7 @@ jobs:
         run: python -m pip install --upgrade twine
 
       - name: Build built distribution
-        uses: pypa/cibuildwheel@v2.14.1
+        uses: pypa/cibuildwheel@v2.15.0
         with:
           package-dir: .
           output-dir: dist/
