### 2023-08-14 162d7d2
Update config after RTD deprecation changes
