### 2023-08-24 78ac923
Update NERSC compilation
