+++ src/triumvirate/resources/params_template.ini
@@ -1,26 +1,37 @@
+# @file params_template.ini
+# @brief Template parameter file for the Triumvirate C++ program.
+# @seealso https://triumvirate.readthedocs.org
+
+
 # -- I/O -----------------------------------------------------------------
 
-# Directories for input/output.
+# Directories for input catalogue(s) and output measurement(s).
 # The paths can be either absolute or relative to the working directory.
 # If unset, the current working directory is assumed.
 catalogue_dir =
 measurement_dir =
 
-# Filenames (with extensions) of input/output sources. [mandatory, optional]
+# Filenames (with extension) of input catalogues.
 data_catalogue_file =
 rand_catalogue_file =
 
-# Column names (comma-separated without space) in catalogue data.
-# [mandatory]
+# Field names of catalogue data columns as a comma-separated list without
+# space in the order of appearance.  Only data columns with the following
+# field names are read from the input catalogue(s),
+#   {'x', 'y', 'z', 'nz', 'ws', 'wc'},
+# where 'x', 'y', 'z' are the Cartesian coordinates, 'nz' is the
+# redshift-dependent number density, 'ws' is the total sample weight
+# (including e.g. completeness weights), and 'wc' is the total clustering
+# weight (including e.g. optimality weights). [mandatory]
 catalogue_columns =
 
-# Tags to be substituted into input/output paths.
+# Tags to be appended as an input/output filename suffix.
 output_tag =
 
 
 # -- Mesh sampling -------------------------------------------------------
 
-# Box size in each dimension (in Mpc/h). [mandatory]
+# Box size in each dimension (in length units).
 boxsize_x =
 boxsize_y =
 boxsize_z =
@@ -31,12 +42,17 @@ ngrid_y =
 ngrid_z =
 
 # Box alignment: {'centre' (default), 'pad'}.
+# The catalogues are either centred or padded from the
+# mininum-coordinate corner.
 alignment = centre
 
 # Padding scale: {'box' (default), 'grid'}.
+# The padding scale is either the box size or the grid cell size.
+# Only applicable if `alignment` is set to 'pad'.
 padscale = box
 
 # Padding factor (as a multiple of the size of padding scale).
+# Only applicable if `alignment` is set to 'pad'.
 padfactor =
 
 # Mesh assignment scheme: {'ngp', 'cic', 'tsc' (default), 'pcs'}.
@@ -49,7 +65,9 @@ interlace = false
 
 # -- Measurements --------------------------------------------------------
 
-# Type of catalogue(s): {'survey', 'random', 'sim'}. [mandatory]
+# Type of catalogue(s): {'survey', 'random', 'sim'}.
+# If of type 'sim', global-plane-parallel measurements are made;
+# otherwise, local-plane-parallel measurements are made. [mandatory]
 catalogue_type =
 
 # Type of measurement:
