+++ .github/workflows/publication.yml
@@ -21,7 +21,7 @@ jobs:
 
     steps:
       - name: Checkout
-        uses: actions/checkout@v3
+        uses: actions/checkout@v4
 
       # By default, the source .md and product .pdf are both named 'paper'
       # and reside in the same directory.
