### 2023-09-08 ca2c6cd
Enhance through thread-local, split spherical Bessel calculators and additional OMP regions
