+++ src/triumvirate/src/field.cpp
@@ -121,7 +121,7 @@ const fftw_complex& MeshField::operator[](int gid) {return this->field[gid];}
 // Mesh grid properties
 // -----------------------------------------------------------------------
 
-long long MeshField::get_grid_index(int i, int j, int k) {
+long long MeshField::ret_grid_index(int i, int j, int k) {
   long long idx_grid =
     (i * this->params.ngrid[1] + j) * this->params.ngrid[2] + k;
   return idx_grid;
@@ -234,7 +234,7 @@ void MeshField::assign_weighted_field_to_mesh_ngp(
     for (int iloc = 0; iloc < order; iloc++) {
       for (int jloc = 0; jloc < order; jloc++) {
         for (int kloc = 0; kloc < order; kloc++) {
-          gid = this->get_grid_index(
+          gid = this->ret_grid_index(
             ijk[iloc][0], ijk[jloc][1], ijk[kloc][2]
           );
           if (0 <= gid && gid < this->params.nmesh) {
@@ -283,7 +283,7 @@ OMP_ATOMIC
       for (int iloc = 0; iloc < order; iloc++) {
         for (int jloc = 0; jloc < order; jloc++) {
           for (int kloc = 0; kloc < order; kloc++) {
-            gid = this->get_grid_index(
+            gid = this->ret_grid_index(
               ijk[iloc][0], ijk[jloc][1], ijk[kloc][2]
             );
             if (0 <= gid && gid < this->params.nmesh) {
@@ -346,7 +346,7 @@ void MeshField::assign_weighted_field_to_mesh_cic(
     for (int iloc = 0; iloc < order; iloc++) {
       for (int jloc = 0; jloc < order; jloc++) {
         for (int kloc = 0; kloc < order; kloc++) {
-          gid = this->get_grid_index(
+          gid = this->ret_grid_index(
             ijk[iloc][0], ijk[jloc][1], ijk[kloc][2]
           );
           if (0 <= gid && gid < this->params.nmesh) {
@@ -395,7 +395,7 @@ OMP_ATOMIC
       for (int iloc = 0; iloc < order; iloc++) {
         for (int jloc = 0; jloc < order; jloc++) {
           for (int kloc = 0; kloc < order; kloc++) {
-            gid = this->get_grid_index(
+            gid = this->ret_grid_index(
               ijk[iloc][0], ijk[jloc][1], ijk[kloc][2]
             );
             if (0 <= gid && gid < this->params.nmesh) {
@@ -475,7 +475,7 @@ void MeshField::assign_weighted_field_to_mesh_tsc(
     for (int iloc = 0; iloc < order; iloc++) {
       for (int jloc = 0; jloc < order; jloc++) {
         for (int kloc = 0; kloc < order; kloc++) {
-          gid = this->get_grid_index(
+          gid = this->ret_grid_index(
             ijk[iloc][0], ijk[jloc][1], ijk[kloc][2]
           );
           if (0 <= gid && gid < this->params.nmesh) {
@@ -543,7 +543,7 @@ OMP_ATOMIC
       for (int iloc = 0; iloc < order; iloc++) {
         for (int jloc = 0; jloc < order; jloc++) {
           for (int kloc = 0; kloc < order; kloc++) {
-            gid = this->get_grid_index(
+            gid = this->ret_grid_index(
               ijk[iloc][0], ijk[jloc][1], ijk[kloc][2]
             );
             if (0 <= gid && gid < this->params.nmesh) {
@@ -613,7 +613,7 @@ void MeshField::assign_weighted_field_to_mesh_pcs(
[…]
