### 2023-09-12 e784043
Refactor assignment scheme conditions
