+++ src/triumvirate/src/field.cpp
@@ -127,6 +127,12 @@ long long MeshField::ret_grid_index(int i, int j, int k) {
   return idx_grid;
 }
 
+void MeshField::shift_grid_indices_fourier(int& i, int& j, int& k) {
+  i = (i < this->params.ngrid[0]/2) ? i : i - this->params.ngrid[0];
+  j = (j < this->params.ngrid[1]/2) ? j : j - this->params.ngrid[1];
+  k = (k < this->params.ngrid[2]/2) ? k : k - this->params.ngrid[2];
+}
+
 void MeshField::get_grid_pos_vector(int i, int j, int k, double rvec[3]) {
   rvec[0] = (i < this->params.ngrid[0]/2) ?
     i * this->dr[0] : (i - this->params.ngrid[0]) * this->dr[0];
@@ -688,26 +694,10 @@ OMP_ATOMIC
   }
 }
 
-double MeshField::calc_assignment_window_in_fourier(int i, int j, int k) {
-  int order = 0;
-  if (this->params.assignment == "ngp") {
-    order = 1;
-  } else
-  if (this->params.assignment == "cic") {
-    order = 2;
-  } else
-  if (this->params.assignment == "tsc") {
-    order = 3;
-  } else
-  if (this->params.assignment == "pcs") {
-    order = 4;
-  } else {
-    order = 0;
-  }
-
-  i = (i < this->params.ngrid[0]/2) ? i : i - this->params.ngrid[0];
-  j = (j < this->params.ngrid[1]/2) ? j : j - this->params.ngrid[1];
-  k = (k < this->params.ngrid[2]/2) ? k : k - this->params.ngrid[2];
+double MeshField::calc_assignment_window_in_fourier(
+  int i, int j, int k, int order=0
+) {
+  this->shift_grid_indices_fourier(i, j, k);
 
   double u_x = M_PI * i / double(this->params.ngrid[0]);
   double u_y = M_PI * j / double(this->params.ngrid[1]);
@@ -1184,7 +1174,9 @@ void MeshField::apply_assignment_compensation() {
       for (int k = 0; k < this->params.ngrid[2]; k++) {
         long long idx_grid = this->ret_grid_index(i, j, k);
 
-        double win = this->calc_assignment_window_in_fourier(i, j, k);
+        double win = this->calc_assignment_window_in_fourier(
+          i, j, k, this->params.assignment_order
+        );
 
         this->field[idx_grid][0] /= win;
         this->field[idx_grid][1] /= win;
@@ -1231,7 +1223,9 @@ void MeshField::inv_fourier_transform_ylm_wgtd_field_band_limited(
           );
 
           // Apply assignment compensation.
-          double win = this->calc_assignment_window_in_fourier(i, j, k);
+          double win = this->calc_assignment_window_in_fourier(
+            i, j, k, this->params.assignment_order
+          );
           fk /= win;
 
           // Weight the field.
@@ -1309,7 +1303,9 @@ void MeshField::inv_fourier_transform_sjl_ylm_wgtd_field(
           field_fourier[idx_grid][0], field_fourier[idx_grid][1]
         );
 
-        double win = this->calc_assignment_window_in_fourier(i, j, k);
+        double win = this->calc_assignment_window_in_fourier(
+          i, j, k, this->params.assignment_order
+        );
         fk /= win;
 
         // Weight the field including the volume normalisation,
@@ -1503,6 +1499,9 @@ void FieldStats::compute_ylm_wgtd_2pt_stats_in_fourier(
     field_a.get_grid_wavevector(i, j, k, kvec);
   };
 
[…]
