### 2023-09-13 3e6268b
Update parameter set and tests on Python side
