### 2023-09-13 10b998f
Refactor interlacing aliasing corrections
