### 2023-09-13 dbfa209
Extend parameter definition for 3-point `form` and `idx_bin`
