+++ src/triumvirate/resources/params_template.yml
@@ -87,7 +87,8 @@ wa_orders:
   i:
   j:
 
-# Form of three-point statistic measurements: {'full', 'diag' (default)}.
+# Form of three-point statistic measurements:
+# {'full', 'diag' (default), 'off-diag', 'row'}.
 form: diag
 
 # Normalisation convention: {
@@ -108,7 +109,10 @@ range: [~, ~]
 # Must be >=2, or >=7 if padding is used in binning.
 num_bins:
 
-# Fixed bin index in the full (2-d) three-point statistics measurements.
+# Fixed bin index when the `form` of three-point statistics measurements
+# is set to 'off-diag' or 'row'.  If 'off-diag', the bin index is the
+# positive off-diagonal index; if `form` is set to 'row', the bin index
+# is the row index.
 idx_bin:
 
 
