### 2023-09-14 359f35d
Add vector counting from binning
