+++ src/triumvirate/src/dataobjs.cpp
@@ -248,6 +248,11 @@ void Binning::compute_binning() {
 
 
 // ***********************************************************************
+// Mesh grids
+// ***********************************************************************
+
+
+// ***********************************************************************
 // Line of sight
 // ***********************************************************************
 
