### 2023-09-14 01133b3
Add range check
