### 2023-09-15 3b89311
Update syntax and add comments
