+++ tests/test_input/params/test_params.yml
@@ -2,44 +2,55 @@
 
 # -- I/O -----------------------------------------------------------------
 
-# Directories for input/output.
+# Directories for input catalogue(s) and output measurement(s).
 # The paths can be either absolute or relative to the working directory.
 # If unset, the current working directory is assumed.
 directories:
   catalogues: tests/test_input/ctlgs
   measurements: tests/test_output/
 
-# Filenames (with extensions) for input/output sources.
+# Filenames (with extension) of input catalogues.  These are relative
+# to the catalogue directory.
 files:
   data_catalogue:
   rand_catalogue:
 
-# Tags to be substituted into input/output paths.
+# Field names of catalogue data columns as a list in the order of
+# appearance.  If unset, the catalogue files are assumed to contain
+# an approriate header.
+catalogue_columns: []
+
+# Tags to be appended as an input/output filename suffix.
 tags:
   output:
 
 
 # -- Sampling ------------------------------------------------------------
 
-# Box size in each dimension (in Mpc/h). [mandatory]
+# Box size in each dimension (in length units).
 boxsize:
   x: 1000.
   y: 1000.
   z: 1000.
 
-# Grid cell number in each dimension. [mandatory]
+# Grid cell number in each dimension.
 ngrid:
   x: 64
   y: 64
   z: 64
 
 # Box alignment: {'centre' (default), 'pad'}.
+# The catalogues are either centred or padded from the
+# mininum-coordinate corner.
 alignment: centre
 
 # Padding scale: {'box' (default), 'grid'}.
+# The padding scale is either the box size or the grid cell size.
+# Only applicable if `alignment` is set to 'pad'.
 padscale: box
 
-# Padding factor (as a multiple of the size of padding scale).
+# Padding factor as a multiple of the size of padding scale.
+# Only applicable if `alignment` is set to 'pad'.
 padfactor: ~
 
 # Mesh assignment scheme: {'ngp', 'cic', 'tsc' (default), 'pcs'}.
@@ -52,15 +63,19 @@ interlace: off
 
 # -- Measurements --------------------------------------------------------
 
-# Type of catalogue(s): {'survey', 'random', 'sim'}. [mandatory]
+# Type of catalogue(s): {'survey', 'random', 'sim', 'none'}.
+# If of type 'sim', global-plane-parallel measurements are made;
+# otherwise, local-plane-parallel measurements are made.
 catalogue_type:
 
-# Type of measurement:
-# {'powspec', '2pcf', '2pcf-win', 'bispec', '3pcf', '3pcf-win', '3pcf-win-wa'}.
-# [mandatory]
+# Type of measurement: {
+#   'powspec', '2pcf', '2pcf-win',
+#   'bispec', '3pcf', '3pcf-win', '3pcf-win-wa',
+#   'modes', 'seps'
+# }. [mandatory]
 statistic_type:
 
-# Degrees of the multipoles. [optional, optional, mandatory]
+# Degrees of the multipoles.
 degrees:
   ell1:
   ell2:
@@ -71,7 +86,8 @@ wa_orders:
   i:
   j:
 
-# Form of the three-point statistic measurements: {'full', 'diag' (default)}.
+# Form of the three-point statistic measurements:
+# {'full', 'diag' (default), 'off-diag', 'row'}.
 form: diag
 
 # Normalisation convention: {
@@ -83,21 +99,29 @@ norm_convention: particle
 # Binning scheme: {'lin' (default), 'log', 'linpad', 'logpad', 'custom'}.
 binning: lin
 
-# Minimum and maximum of the binning range (i.e. range of measurement scales)
-# as a list.
-# The binning coordinate is either wavenumbers (in h/Mpc) in Fourier space,
-# or separations (in Mpc/h) in configuration space. [mandatory]
[…]
