+++ src/triumvirate/resources/params_template.ini
@@ -101,8 +101,8 @@ norm_convention = particle
 binning = lin
 
 # Minimum and maximum of the range of measurement scales.
-# The binning coordinate is either wavenumbers (in h/Mpc) in Fourier space,
-# or separations (in Mpc/h) in configuration space. [mandatory]
+# The binning coordinate is either wavenumbers in Fourier space,
+# or separations in configuration space. [mandatory]
 bin_min =
 bin_max =
 
