+++ src/triumvirate/resources/params_template.yml
@@ -105,8 +105,8 @@ norm_convention: particle
 binning: lin
 
 # Range of measurement scales.
-# The binning coordinate is either wavenumbers (in h/Mpc) in Fourier space,
-# or separations (in Mpc/h) in configuration space.
+# The binning coordinate is either wavenumbers in Fourier space,
+# or separations in configuration space.
 range: [~, ~]
 
 # Number of measurement bins (i.e. data vector dimension).
