### 2023-09-19 225129e
Add wrapper for `BinnedVectors`. Remove blank line
