+++ src/triumvirate/src/parameters.cpp
@@ -808,13 +808,13 @@ int ParameterSet::validate() {
     && this->save_binned_vectors == ""
   ) {
     this->save_binned_vectors = default_bvec_sfilepath;  // transmutation
-  }
-  if (trvs::currTask == 0) {
-    trvs::logger.warn(
-      "`save_binned_vectors` is overriden, as `statistic_type` is '%s' "
-      "so binned vectors are saved as the output to the default path.",
-      this->statistic_type.c_str()
-    );
+    if (trvs::currTask == 0) {
+      trvs::logger.warn(
+        "`save_binned_vectors` is overriden, as `statistic_type` is '%s' "
+        "so binned vectors are saved as the output to the default path.",
+        this->statistic_type.c_str()
+      );
+    }
   }
 
   if (trvs::currTask == 0) {
