+++ src/triumvirate/resources/params_template.ini
@@ -23,7 +23,7 @@ rand_catalogue_file =
 # where 'x', 'y', 'z' are the Cartesian coordinates, 'nz' is the
 # redshift-dependent number density, 'ws' is the total sample weight
 # (including e.g. completeness weights), and 'wc' is the total clustering
-# weight (including e.g. optimality weights). [mandatory]
+# weight (including e.g. optimality weights).
 catalogue_columns =
 
 # Tags to be appended as an input/output filename suffix.
@@ -78,7 +78,7 @@ catalogue_type =
 # }. [mandatory]
 statistic_type =
 
-# Degrees of the multipoles. [optional, optional, mandatory]
+# Degrees of the multipoles.
 ell1 =
 ell2 =
 ELL =
