+++ src/triumvirate/resources/params_template.yml
@@ -1,9 +1,8 @@
----
-
 # @file params_template.yml
 # @brief Template parameter file for the Triumvirate Python package.
 # @seealso https://triumvirate.readthedocs.org
 
+---
 
 # -- I/O -----------------------------------------------------------------
 
