### 2023-09-20 b765259
Update `dataobjs` structs and refactor
