+++ src/triumvirate/src/io.cpp
@@ -356,15 +356,15 @@ void print_measurement_datatab_to_file(
   );
 
   // Print data table.
-  for (int ibin = 0; ibin < params.num_bins; ibin++) {
+  for (int idx_dv = 0; idx_dv < meas_powspec.dim; idx_dv++) {
     std::fprintf(
       fileptr,
       "%.9e\t%.9e\t%10d\t% .9e\t% .9e\t% .9e\t% .9e\n",
-      meas_powspec.kbin[ibin],
-      meas_powspec.keff[ibin],
-      meas_powspec.nmodes[ibin],
-      meas_powspec.pk_raw[ibin].real(), meas_powspec.pk_raw[ibin].imag(),
-      meas_powspec.pk_shot[ibin].real(), meas_powspec.pk_shot[ibin].imag()
+      meas_powspec.kbin[idx_dv],
+      meas_powspec.keff[idx_dv],
+      meas_powspec.nmodes[idx_dv],
+      meas_powspec.pk_raw[idx_dv].real(), meas_powspec.pk_raw[idx_dv].imag(),
+      meas_powspec.pk_shot[idx_dv].real(), meas_powspec.pk_shot[idx_dv].imag()
     );
   }
 }
@@ -382,14 +382,14 @@ void print_measurement_datatab_to_file(
   );
 
   // Print data table.
-  for (int ibin = 0; ibin < params.num_bins; ibin++) {
+  for (int idx_dv = 0; idx_dv < meas_2pcf.dim; idx_dv++) {
     std::fprintf(
       fileptr,
       "%.9e\t%.9e\t%10d\t% .9e\t% .9e\n",
-      meas_2pcf.rbin[ibin],
-      meas_2pcf.reff[ibin],
-      meas_2pcf.npairs[ibin],
-      meas_2pcf.xi[ibin].real(), meas_2pcf.xi[ibin].imag()
+      meas_2pcf.rbin[idx_dv],
+      meas_2pcf.reff[idx_dv],
+      meas_2pcf.npairs[idx_dv],
+      meas_2pcf.xi[idx_dv].real(), meas_2pcf.xi[idx_dv].imag()
     );
   }
 }
@@ -407,14 +407,14 @@ void print_measurement_datatab_to_file(
   );
 
   // Print data table.
-  for (int ibin = 0; ibin < params.num_bins; ibin++) {
+  for (int idx_dv = 0; idx_dv < meas_2pcf_win.dim; idx_dv++) {
     std::fprintf(
       fileptr,
       "%.9e\t%.9e\t%10d\t% .9e\t% .9e\n",
-      meas_2pcf_win.rbin[ibin],
-      meas_2pcf_win.reff[ibin],
-      meas_2pcf_win.npairs[ibin],
-      meas_2pcf_win.xi[ibin].real(), meas_2pcf_win.xi[ibin].imag()
+      meas_2pcf_win.rbin[idx_dv],
+      meas_2pcf_win.reff[idx_dv],
+      meas_2pcf_win.npairs[idx_dv],
+      meas_2pcf_win.xi[idx_dv].real(), meas_2pcf_win.xi[idx_dv].imag()
     );
   }
 };
@@ -438,23 +438,25 @@ void print_measurement_datatab_to_file(
   std::fprintf(
     fileptr,
     "%s "
-    "[0] k1_cen, [1] k1_eff, [2] k2_cen, [3] k2_eff, [4] nmodes, "
[…]
