### 2023-09-25 a3b9ed8
Refactor binning for 3-point statistics
