+++ src/triumvirate/main/triumvirate.cpp
@@ -558,7 +558,7 @@ int main(int argc, char* argv[]) {
     } else
     if (params.form == "row") {
       std::snprintf(
-        save_filepath, sizeof(save_filepath), "%s/bk%d%d%d_bin%d%s",
+        save_filepath, sizeof(save_filepath), "%s/bk%d%d%d_row%d%s",
         params.measurement_dir.c_str(),
         params.ell1, params.ell2, params.ELL, params.idx_bin,
         params.output_tag.c_str()
@@ -593,25 +593,18 @@ int main(int argc, char* argv[]) {
     std::fclose(save_fileptr);
   } else
   if (params.statistic_type == "3pcf") {
-    if (params.form == "full") {
+    if (params.form == "full" || params.form == "diag") {
       std::snprintf(
-        save_filepath, sizeof(save_filepath), "%s/zeta%d%d%d_%s",
+        save_filepath, sizeof(save_filepath), "%s/zeta%d%d%d_%s%s",
         params.measurement_dir.c_str(),
         params.ell1, params.ell2, params.ELL,
-        params.output_tag.c_str()
-      );
-    } else
-    if (params.form == "diag") {
-      std::snprintf(
-        save_filepath, sizeof(save_filepath), "%s/zeta%d%d%d_diag%s",
-        params.measurement_dir.c_str(),
-        params.ell1, params.ell2, params.ELL,
+        params.form.c_str(),
         params.output_tag.c_str()
       );
     } else
     if (params.form == "off-diag") {
       std::snprintf(
-        save_filepath, sizeof(save_filepath), "%s/zeta%d%d%d_off-diag%d%s",
+        save_filepath, sizeof(save_filepath), "%s/zeta%d%d%d_offdiag%d%s",
         params.measurement_dir.c_str(),
         params.ell1, params.ell2, params.ELL, params.idx_bin,
         params.output_tag.c_str()
@@ -619,7 +612,7 @@ int main(int argc, char* argv[]) {
     } else
     if (params.form == "row") {
       std::snprintf(
-        save_filepath, sizeof(save_filepath), "%s/zeta%d%d%d_bin%d%s",
+        save_filepath, sizeof(save_filepath), "%s/zeta%d%d%d_row%d%s",
         params.measurement_dir.c_str(),
         params.ell1, params.ell2, params.ELL, params.idx_bin,
         params.output_tag.c_str()
@@ -654,19 +647,12 @@ int main(int argc, char* argv[]) {
     std::fclose(save_fileptr);
   } else
   if (params.statistic_type == "3pcf-win") {
-    if (params.form == "full") {
+    if (params.form == "full" || params.form == "diag") {
       std::snprintf(
-        save_filepath, sizeof(save_filepath), "%s/zetaw%d%d%d_%s",
+        save_filepath, sizeof(save_filepath), "%s/zetaw%d%d%d_%s%s",
         params.measurement_dir.c_str(),
         params.ell1, params.ell2, params.ELL,
-        params.output_tag.c_str()
-      );
-    } else
-    if (params.form == "diag") {
-      std::snprintf(
-        save_filepath, sizeof(save_filepath), "%s/zetaw%d%d%d_diag%s",
-        params.measurement_dir.c_str(),
-        params.ell1, params.ell2, params.ELL,
+        params.form.c_str(),
         params.output_tag.c_str()
       );
     } else
@@ -680,7 +666,7 @@ int main(int argc, char* argv[]) {
     } else
     if (params.form == "row") {
       std::snprintf(
-        save_filepath, sizeof(save_filepath), "%s/zetaw%d%d%d_bin%d%s",
[…]
