+++ src/triumvirate/src/threept.cpp
@@ -312,6 +312,7 @@ trv::BispecMeasurements compute_bispec(
   trvm::SphericalBesselCalculator sj_b(params.ell2);  // j_l_b
 
   // Compute bispectrum terms including shot noise.
+  int count_terms = 0;
   for (int m1_ = - params.ell1; m1_ <= params.ell1; m1_++) {
     for (int m2_ = - params.ell2; m2_ <= params.ell2; m2_++) {
       // Check for if all Wigner-3j symbols are zero.
@@ -394,8 +395,7 @@ trv::BispecMeasurements compute_bispec(
               dn_00, ylm_k_b, k_lower, k_upper, k_eff_b_, nmodes_b_
             );
 
-            // Only record the binned coordinates and counts once.
-            if (M_ == 0 && m1_ == 0 && m2_ == 0) {
+            if (count_terms == 0) {
               k1bin_dv[idx_dv] = kbinning.bin_centres[ibin];
               k2bin_dv[idx_dv] = kbinning.bin_centres[ibin];
               k1eff_dv[idx_dv] = k_eff_a_;
@@ -451,8 +451,7 @@ trv::BispecMeasurements compute_bispec(
               dn_00, ylm_k_b, k_lower_b, k_upper_b, k_eff_b_, nmodes_b_
             );
 
-            // Only record the binned coordinates and counts once.
-            if (M_ == 0 && m1_ == 0 && m2_ == 0) {
+            if (count_terms == 0) {
               k1bin_dv[idx_dv] = kbinning.bin_centres[ibin_row];
               k2bin_dv[idx_dv] = kbinning.bin_centres[ibin_col];
               k1eff_dv[idx_dv] = k_eff_a_;
@@ -514,8 +513,7 @@ trv::BispecMeasurements compute_bispec(
               dn_00, ylm_k_b, k_lower_b, k_upper_b, k_eff_b_, nmodes_b_
             );
 
-            // Only record the binned coordinates and counts once.
-            if (M_ == 0 && m1_ == 0 && m2_ == 0) {
+            if (count_terms == 0) {
               k1bin_dv[idx_dv] = kbinning.bin_centres[ibin_row];
               k2bin_dv[idx_dv] = kbinning.bin_centres[ibin_col];
               k1eff_dv[idx_dv] = k_eff_a_;
@@ -572,8 +570,7 @@ trv::BispecMeasurements compute_bispec(
                 dn_00, ylm_k_b, k_lower_b, k_upper_b, k_eff_b_, nmodes_b_
               );
 
-              // Only record the binned coordinates and counts once.
-              if (M_ == 0 && m1_ == 0 && m2_ == 0) {
+              if (count_terms == 0) {
                 k1bin_dv[idx_dv] = kbinning.bin_centres[idx_row];
                 k2bin_dv[idx_dv] = kbinning.bin_centres[idx_col];
                 k1eff_dv[idx_dv] = k_eff_a_;
@@ -753,6 +750,7 @@ trv::BispecMeasurements compute_bispec(
           sn_dv[idx_dv] += coupling * S_ij_k;
         }
 
+        count_terms++;
         if (trvs::currTask == 0) {
           trvs::logger.stat(
             "Bispectrum term at orders (m1, m2, M) = (%d, %d, %d) computed.",
@@ -901,6 +899,7 @@ trv::ThreePCFMeasurements compute_3pcf(
   trvm::SphericalBesselCalculator sj_b(params.ell2);  // j_l_b
 
   // Compute 3PCF terms including shot noise.
+  int count_terms = 0;
   for (int m1_ = - params.ell1; m1_ <= params.ell1; m1_++) {
[…]
