### 2023-09-27 ba30bc8
Add Intel compiler options
