+++ src/triumvirate/threept.py
@@ -1199,7 +1199,7 @@ def _compute_3pt_stats_sim_like(threept_algofunc, catalogue_data,
         if logger:
             logger.info(
                 "Inserted missing 'nz' field "
-                "based on particle count and boxsize."
+                "based on particle count and box size."
             )
 
     if logger:
