+++ src/triumvirate/_field.pyx
@@ -1,5 +1,5 @@
 """
-Mesh fields and statistics (:mod:`~triumvirate._field`)
+Mesh field and statistics (:mod:`~triumvirate._field`)
 ==========================================================================
 
 Parse mesh fields and their statistics.
