### 2023-10-01 #5571 @arfon
Friendly reminder to get to these updates soon please @MikeSWang. @alfonso-veropalumbo – I think the author asked you a question here too...

> A quick question about comment 1 from @alfonso-veropalumbo: we have included some performance metrics for the bispectrum, so I guess you are asking about the 3PCF in configuration space in particular?
