### 2023-10-02 67e7d3d
Update core and API docs
