+++ src/triumvirate/src/monitor.cpp
@@ -332,7 +332,25 @@ void display_prog_notice() {
     "  //      \\                                                         \n"
     " //________\\    (C) 2023 Mike S Wang & Naonori Sugiyama [GPLv3]     \n"
     "                                                                     \n"
-    "  •••  Three-Point Clustering Measurements in LSS  •••             \n\n"
+    " •••  Three-Point Clustering Measurements in LSS  •••              \n\n"
+  );
+}
+
+void display_prog_licence() {
+  std::printf(
+    "This program is free software: you can redistribute it and/or modify \n"
+    "it under the terms of the GNU General Public License as published by \n"
+    "the Free Software Foundation, either version 3 of the License, or    \n"
+    "(at your option) any later version.                                  \n"
+    "                                                                     \n"
+    "This program is distributed in the hope that it will be useful, but  \n"
+    "WITHOUT ANY WARRANTY; without even the implied warranty of           \n"
+    "MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU     \n"
+    "General Public License for more details.                             \n"
+    "                                                                     \n"
+    "You should have received a copy of the GNU General Public License    \n"
+    "along with this program. If not, see <https://www.gnu.org/licenses/>.\n"
+    "                                                                     \n"
   );
 }
 
