### 2023-10-04 #5571 @alfonso-veropalumbo
> Friendly reminder to get to these updates soon please @MikeSWang. @alfonso-veropalumbo – I think the author asked you a question here too...
>
> > A quick question about comment 1 from @alfonso-veropalumbo: we have included some performance metrics for the bispectrum, so I guess you are asking about the 3PCF in configuration space in particular?

Yes, sure, I think it could be useful. Computational time is the main issue, computationally speaking, for configuration space estimators (and particularly for 3PCF).
