### 2023-10-04 4ea795b
Update developer's and release notes. Bump up version
