+++ NOTES.md
@@ -1,7 +1,25 @@
 # Developer's Notes
 
+## Pre-release checklist
+
+- [ ] Bump fallback version number;
+- [ ] Check agreement amongst read-me pages:
+    - [repo read-me](README.md)
+    - [PyPI read-me](README.rst)
+    - [ReadtheDocs homepage](docs/source/index.rst)
+- [ ] Check agreement between installation instructions:
+    - [repo read-me](README.md#Installation)
+    - [PyPI read-me](README.rst#Installation)
+    - [ReadtheDocs](docs/source/installation.rst)
+- [ ] Check agreement between release notes:
+    - [change-log](CHANGELOG.md)
+    - [release history](docs/source/releases.rst)
+
+
+## Code tags
+
 - The following code tags are in use:
-    - 'RFE';
+    - 'RFE', 'FUTURE';
     - 'CAVEAT';
     - 'SEE';
     - 'STYLE'.
