+++ .github/workflows/cd.yml
@@ -110,7 +110,7 @@ jobs:
         run: python -m pip install --upgrade twine
 
       - name: Build built distribution
-        uses: pypa/cibuildwheel@v2.16.1
+        uses: pypa/cibuildwheel@v2.16.2
         with:
           package-dir: .
           output-dir: dist/
