+++ .github/workflows/cd_xp.yml
@@ -49,7 +49,7 @@ jobs:
 
     runs-on: ${{ matrix.os }}
 
-    timeout-minutes: 120
+    timeout-minutes: 180
 
     # Allow manual trigger from inputs
     # (skip automatic trigger with ``&& false``).
@@ -58,8 +58,27 @@ jobs:
       github.event.inputs.run_build_bdist_xplat == 'true'
 
     steps:
-      - name: Checkout
+      - name: Checkout (automatic trigger)
+        if: github.event_name != 'workflow_dispatch'
         uses: actions/checkout@v4
+        with:
+          fetch-depth: 0
+
+      - name: Checkout (manual trigger)
+        if: >
+          github.event.inputs.run_build_bdist_xplat == 'true' &&
+          github.event.inputs.version_tag == ''
+        uses: actions/checkout@v3
+        with:
+          fetch-tags: true
+
+      - name: Checkout (manual trigger with tag)
+        if: >
+          github.event.inputs.run_build_bdist_xplat == 'true' &&
+          github.event.inputs.version_tag != ''
+        uses: actions/checkout@v3
+        with:
+          ref: refs/tags/${{ github.event.inputs.version_tag }}
 
       # Use QEMU for non-native Linux runner.
       - name: Set up QEMU (Linux)
