+++ src/triumvirate/resources/params_template.ini
@@ -1,6 +1,6 @@
 # @file params_template.ini
 # @brief Template parameter file for the Triumvirate C++ program.
-# @seealso https://triumvirate.readthedocs.org
+# @seealso https://triumvirate.readthedocs.io
 
 
 # -- I/O -----------------------------------------------------------------
