+++ src/triumvirate/resources/params_template.yml
@@ -1,6 +1,6 @@
 # @file params_template.yml
 # @brief Template parameter file for the Triumvirate Python package.
-# @seealso https://triumvirate.readthedocs.org
+# @seealso https://triumvirate.readthedocs.io
 
 ---
 
