+++ publication/joss/paper.md
@@ -51,7 +51,7 @@ complementarity, methods for measuring two-point clustering statistics are
 also included in the package.
 
 
-[^1]: [github.com/naonori/hitomi/](https://github.com/naonori/hitomi/)
+[^1]: [github.com/naonori/hitomi](https://github.com/naonori/hitomi)
 
 
 # Statement of need
@@ -98,20 +98,61 @@ can compute:
     models derived in Fourier space through the Hankel transform
     [@Wilson:2016;@Sugiyama:2019].
 
+For the global plane-parallel estimators, the simulation box is placed at
+the spatial infinity (or equivalently the observer is), so that the
+line of sight to each particle can be treated as the same and taken to be
+along the $z$-axis. For the local plane-parallel estimators, the observer
+is placed at the origin in the survey coordinates, and the line of sight
+is chosen to point towards one of the particles in a triplet or pair for
+three- or two-point clustering measurements respectively.
+
+The geometry of the survey leaves an imprint on the clustering statistics,
+where in Fourier space the effect is a convolution with the survey window
+function. This convolution mixes different multipoles of the underlying
+clustering statistics and the survey window, and the precise convolution
+formula (i.e. the number of multipoles to include in modelling) needed to
+achieve a given level of convergence depends on the precise survey geometry
+including any sample weights applied. Therefore the functionality to
+measure the window function is an integral part of this program.
+
 These functionalities are essential to cosmological inference pipelines,
 and can help validate any analytical covariance matrix predictions against
 sample estimates. Since precise covariance matrix estimates usually
 require clustering measurements repeated over a large number of simulated
 mock catalogues, computational efficiency is an important objective.
-Finally,  `Triumvirate` also enables comparison studies between
-alternative compressed statistics of three-point clustering
-[e.g. @Scoccimarro:2015; @Slepian:2018], which may have different
-constraining power on different cosmological parameters.
+
+Finally, `Triumvirate` also enables comparison studies between
+alternative compressed statistics of three-point clustering, which may have
+different constraining power on different cosmological parameters. There
+are existing software packages for some of these alternative approaches:
+
+  * `pylians`[^4] [@VillaescusaNavarro:2018] computes the bispectrum with
+    the Scoccimarro estimator [@Scoccimarro:2015] for triangle
+    configurations parametrised by two wavenumbers and the angle between
+    the corresponding wavevectors;
+
+  * `nbodykit`[^5] [@Hand:2018] computes the isotropised 3PCF with a
+    pair-counting algorithm [@Slepian:2015], although in principle
+    this can be generalised to anisotropic 3PCF [@Slepian:2018], or be
+    implemented using FFTs [@Slepian:2016];
+
+  * @Philcox:2021[^6] advocates a windowless cubic estimator for the
+    bispectrum in the Soccimarro decomposition, which can be evaluated
+    using FFTs. However, this approach requires the inversion of a Fisher
+    matrix obtained from a suite of Monte Carlo realisations.
+
+As these programs use a different decomposition of three-point clustering
+statistics and focus on either configuration- or Fourier-space statistics
+only, `Triumvirate` fulfills complementary needs in current galaxy
+clustering analyses.
 
 
[…]
