+++ publication/joss/paper.bib
@@ -1,19 +1,19 @@
 @article{Sugiyama:2019,
     author = {Sugiyama, Naonori S. and Saito, Shun and Beutler, Florian and Seo, Hee-Jong},
-	title = {A complete {FFT}-based decomposition formalism for the redshift-space bispectrum},
-	journal = {Mon.~Not.~R.~Astron.~Soc.},
-	year = {2019},
-	volume = {484},
-	number = {1},
-	pages = {364--384},
-	doi = {10.1093/mnras/sty3249},
+    title = {A complete {FFT}-based decomposition formalism for the redshift-space bispectrum},
+    journal = {Mon.~Not.~R.~Astron.~Soc.},
+    year = {2019},
+    volume = {484},
+    number = {1},
+    pages = {364--384},
+    doi = {10.1093/mnras/sty3249},
 }
 
 @article{Sugiyama:2018,
     author = {Sugiyama, Naonori S. and Shiraishi, Maresuke and Okumura, Teppei},
     title = {{Limits on statistical anisotropy from {BOSS DR12} galaxies using bipolar spherical harmonics}},
-	journal = {Mon.~Not.~R.~Astron.~Soc.},
-	year = {2018},
+    journal = {Mon.~Not.~R.~Astron.~Soc.},
+    year = {2018},
     volume = {473},
     number = {2},
     pages = {2737--2752},
@@ -65,13 +65,13 @@
 
 @article{Bernardeau:2002,
     author = {Bernardeau, F. and Colombi, S. and Gazta{\~n}aga, E. and Scoccimarro, R.},
-	title = {Large-scale structure of the Universe and cosmological perturbation theory},
-	journal = {Phys.~Rep.},
-	year = {2002},
-	volume = {367},
-	issues = {1--3},
-	pages = {1--248},
-	doi = {10.1016/s0370-1573(02)00135-7},
+    title = {Large-scale structure of the Universe and cosmological perturbation theory},
+    journal = {Phys.~Rep.},
+    year = {2002},
+    volume = {367},
+    issues = {1--3},
+    pages = {1--248},
+    doi = {10.1016/s0370-1573(02)00135-7},
 }
 
 @article{Sefusatti:2006,
@@ -92,11 +92,11 @@
         {SDSS-III Baryon Oscillation Spectroscopic Survey}:
         cosmological analysis of the {DR}12 galaxy sample
     },
-	journal = {Mon.~Not.~R.~Astron.~Soc.},
+    journal = {Mon.~Not.~R.~Astron.~Soc.},
     year = {2017},
-	volume = {470},
-	number = {3},
-	pages = {2617--2652},
+    volume = {470},
+    number = {3},
+    pages = {2617--2652},
     doi = {10.1093/mnras/stx721},
 }
 
@@ -107,7 +107,7 @@
         {Cosmological} implications from two decades of spectroscopic surveys
         at the {Apache Point Observatory}
     },
-	journal = {Phys.~Rev.~D},
+    journal = {Phys.~Rev.~D},
     year = {2021},
     volume = {103},
     issue = {8},
@@ -117,35 +117,13 @@
 
 @article{Scoccimarro:1999,
     author = {Scoccimarro, Rom\'an and Couchman, H. M. P. and Frieman, Joshua A.},
-	title = {{The Bispectrum as a Signature of Gravitational Instability in Redshift Space}},
-	journal = {Astrophys.~J.},
-	year = {1999},
-	volume = {517},
-	number = {2},
-	pages = {531--540},
-	doi = {10.1086/307220},
-}
-
-@article{Scoccimarro:2015,
-    author = {Scoccimarro, Rom\'an},
-    title = {Fast estimators for redshift-space clustering},
-    journal = {Phys.~Rev.~D},
-    year = {2015},
[…]
