+++ docs/source/releases.rst
@@ -19,7 +19,8 @@ Features
   to a file in C++.
 
 - Add more forms of three-point statistics including the off-diagonal and
-  full form (the original 'full' form is renamed to 'row').
+  full form (the original 'full' form is renamed to 'row')
+  (`#22 <https://github.com/MikeSWang/Triumvirate/issues/22>`_).
 
 Improvements
 ------------
