### 2023-10-09 c187bd8
Fix copyright symbol
