### 2023-10-11 a67ebd0
Refactor gamma function computations
