+++ src/triumvirate/src/fftlog.cpp
@@ -32,56 +32,133 @@ namespace trv {
 
 namespace maths {
 
-double calc_kr_pivot_lowring(
-  double mu, double q, double L, int N, double kr_c
+HankelTransform::HankelTransform(double mu, double q) {
+  this->order = mu;
+  this->bias = q;
+}
+
+void HankelTransform::initialise(
+  std::vector<double> sample_pts, double kr_c, bool lowring
 ) {
-  double x_p = (mu + 1 + q)/2;
-  double x_m = (mu + 1 - q)/2;
-  double y = M_PI * N / (2*L);
+  // Initialise pre-sample points.
+  if (sample_pts.size() < 2) {
+    throw trv::sys::InvalidParameterError(
+      "The number of sample points must be at least 2."
+    );
+  }
 
-  // Note that no minus is involved by complex conjugation of
-  // the gamma function.
-  double lnr_, phi_p, phi_m;
-  trvm::get_lngamma_components(x_p, y, lnr_, phi_p);
-  trvm::get_lngamma_components(x_m, y, lnr_, phi_m);
+  this->pre_sampts = sample_pts;
+  this->nsamp = sample_pts.size();
+  this->logres =
+    std::log(sample_pts.back() / sample_pts.front()) / (this->nsamp - 1);
 
-  double arg_ = std::log(2/kr_c) * N / L + (phi_p + phi_m) / M_PI;
-  double arg_int = std::round(arg_);
-  if (arg_ != arg_int) {
-    kr_c *= std::exp(L / N * (arg_ - arg_int));
+  // Initialise the transform.
+  if (lowring) {
+    this->pivot = this->calc_lowring_pivot(this->logres, kr_c);
+  } else {
+    if (kr_c <= 0.) {
+      throw trv::sys::InvalidParameterError(
+        "Pivot value must be positive."
+      );
+    }
+    this->pivot = kr_c;
+  }
+
+  this->kernel = this->compute_kernel_coeff();
+
+  // Initialise post-sample points.
+  this->post_sampts.resize(this->nsamp);
+  for (int j = 0; j < this->nsamp; j++) {
+    this->post_sampts[j] = this->pivot / this->pre_sampts[this->nsamp - j - 1];
+  }
+
+  // // Alternative to the for-loop block above, note that k_c and r_c are
+  // // both exp(L/2) times `k0` and `r0`.
+  // // STYLE: Standard naming convention is not followed below.
+  // int N = this->nsamp;
+  // double dL = this->logres;
+  // double kr_c_used = this->pivot;
+  // double r0 = this->pre_sampts[0];
+
+  // double kr_0 = kr_c_used * std::exp(- N * dL);
+  // double k0 = kr_0 / r0;
+}
+
+double HankelTransform::calc_lowring_pivot(double delta, double kr_c) {
+  // STYLE: Standard naming convention is not followed below.
+  double mu = this->order;
+  double q = this->bias;
+  double dL = delta;
+
+  double x_p = (mu + 1. + q)/2.;
+  double x_m = (mu + 1. - q)/2.;
+  double y = M_PI / (2.*dL);
+
+  if (kr_c > 0.) {
+    // Note that no minus sign is involved in the phase by
+    // complex conjugation of the gamma function.
+    double lnr_p, lnr_m, phi_p, phi_m;
+    trvm::get_lngamma_parts(x_p, y, lnr_p, phi_p);
+    trvm::get_lngamma_parts(x_m, y, lnr_m, phi_m);
+
+    double argphase = std::log(2./kr_c) / dL + (phi_p + phi_m) / M_PI;
+
[…]
