+++ examples/FFTLog/fftlog_demo.py
@@ -1,0 +1,443 @@
+"""Demonstrate the use of the FFTLog algorithm for Hankel-like transforms.
+
+.. attention::
+
+    Requires Python >=3.11 for the following language features:
+
+    - Match statement (PEP 634);
+    - Self type (PEP 673).
+
+"""
+import argparse
+import os.path as osp
+import sys
+import warnings
+from functools import wraps
+from typing import Callable, Literal, Self, Tuple, Union
+
+import matplotlib.pyplot as plt
+import numpy as np
+import hankl
+from scipy.interpolate import InterpolatedUnivariateSpline
+
+from triumvirate._fftlog import HankelTransform
+from triumvirate.transforms import SphericalBesselTransform
+
+
+def parse_parameters() -> argparse.Namespace:
+    """Initialise the command-line argument parser.
+
+    Returns
+    -------
+    pars
+        Parsed command-line parameters.
+
+    """
+    parser = argparse.ArgumentParser(
+        description="FFTLog Hankel transform demo."
+    )
+
+    parser.add_argument(
+        '-t', "--test-case", type=str,
+        choices=['hankel-sym', 'hankel-asym', 'sj-sym', 'sj-cosmo'],
+        required=True
+    )
+    parser.add_argument(
+        '--order', type=int, default=0,
+        help="Order of the Hankel transform."
+    )
+    parser.add_argument(
+        '--degree', type=int, default=0,
+        help="Degree of the spherical Bessel transform."
+    )
+    parser.add_argument(
+        '--show-diff', action='store_true',
+        help="Show the difference between the transform results."
+    )
+
+    pars = parser.parse_args()
+
+    # HACK: Override the degree to be 0 to match the external samples.
+    if pars.test_case == 'sj-cosmo' and pars.degree != 0:
+        pars.degree = 0
+        warnings.warn(
+            "Spherical Bessel transform degree overriden to 0 "
+            "when external cosmological multipole samples are referenced "
+            "as they are computed for the monopole only."
+        )
+
+    return pars
+
+
+def get_analy_func_pair(
+        testcase: Literal['hankel-sym', 'hankel-asym', 'sj-sym']
+    ) -> Tuple[Tuple[Callable, str], Tuple[Callable, str]]:
+    """Get pre- and post-transform analytical functions for the given
+    test case.
+
+    Parameters
+    ----------
+    testcase
+        Test cases:
+
+        - 'hankel-sym' for symmetric-form functions for the
+          Hankel transform;
+        - 'hankel-asym' for asymmetric-form functions for the
+          Hankel transform;
+        - 'sj-sym' for symmetric-form functions for the spherical
+          Bessel transform.
+
+    Returns
+    -------
+    Pre- and post-transform functions with their TeX strings.
+
+    """
+    match testcase:
+        case 'hankel-sym':
+            return ((
+                lambda r, mu=0.: r**(mu + 1.) * np.exp(-r*r/2.),
+                r"$f(r) = r^{\mu + 1} \mathrm{e}^{-r^2 / 2}$"
+            ), (
+                lambda k, mu=0.: k**(mu + 1.) * np.exp(-k*k/2.),
+                r"$g(k) = k^{\mu + 1} \mathrm{e}^{-k^2 / 2}$"
+            ))
+        case 'hankel-asym':
+            return ((
+                lambda r, **kwargs: r * (1 + r**2)**(-3/2),
[…]
