+++ examples/FFTLog/fftlog_demo.py
@@ -13,15 +13,22 @@ import os.path as osp
 import sys
 import warnings
 from functools import wraps
-from typing import Callable, Literal, Self, Tuple, Union
+from typing import Callable, List, Literal, Self, Tuple, Union
 
 import matplotlib.pyplot as plt
 import numpy as np
-import hankl
 from scipy.interpolate import InterpolatedUnivariateSpline
 
 from triumvirate._fftlog import HankelTransform
 from triumvirate.transforms import SphericalBesselTransform
+
+try:
+    import hankl
+except ImportError:
+    warnings.warn(
+        "Could not import `hankl` package. "
+        "Some of the test cases will not be available."
+    )
 
 
 def parse_parameters() -> argparse.Namespace:
@@ -70,8 +77,8 @@ def parse_parameters() -> argparse.Namespace:
 
 
 def get_analy_func_pair(
-        testcase: Literal['hankel-sym', 'hankel-asym', 'sj-sym']
-    ) -> Tuple[Tuple[Callable, str], Tuple[Callable, str]]:
+    testcase: Literal['hankel-sym', 'hankel-asym', 'sj-sym']
+) -> Tuple[Tuple[Callable, str], Tuple[Callable, str]]:
     """Get pre- and post-transform analytical functions for the given
     test case.
 
@@ -124,10 +131,10 @@ def get_analy_func_pair(
 
 
 def get_external_samples(
-        presamp_file: str = "pk_lgsamps.dat",
-        postsamp_file: str = "xir_postsamps.dat",
-        samp_dir: str = "examples/FFTLog/storage/input/samps"
-    ) -> Tuple[Tuple[np.ndarray, np.ndarray], Tuple[np.ndarray, np.ndarray]]:
+    presamp_file: str = "pk_lgsamps.dat",
+    postsamp_file: str = "xir_postsamps.dat",
+    samp_dir: str = "examples/FFTLog/storage/input/samps"
+) -> Tuple[Tuple[np.ndarray, np.ndarray], Tuple[np.ndarray, np.ndarray]]:
     """Get pre- and post-transform samples from external files.
 
     Parameters
@@ -151,7 +158,7 @@ def get_external_samples(
     return presamples, postsamples
 
 
-class ComparisonPlot:
+class comparison_plot:
     """Comparison plot context manager.
 
     Parameters
@@ -175,15 +182,15 @@ class ComparisonPlot:
     """
 
     def __init__(
-            self,
-            comp_type: Literal['analytic', 'samples'],
-            plot_diff: bool = True,
-            xlim: Union[Tuple[float, float], None] = None,
-            ylim: Union[Tuple[float, float], None] = None,
-            xlabel: Union[str, None] = None,
-            ylabel: Union[str, None] = None,
-            title: Union[str, None] = None
-        ) -> None:
+        self,
+        comp_type: Literal['analytic', 'samples'],
+        plot_diff: bool = True,
+        xlim: Union[Tuple[float, float], None] = None,
+        ylim: Union[Tuple[float, float], None] = None,
+        xlabel: Union[str, None] = None,
+        ylabel: Union[str, None] = None,
+        title: Union[str, None] = None
+    ) -> None:
         # Store figure attributes.
         self._comp_type = comp_type
         self._flag_diff = plot_diff
@@ -236,7 +243,7 @@ class ComparisonPlot:
 
         self._ax_comp.set_ylabel(self._ylabel)
         if self._flag_diff:
-             self._ax_diff.set_ylabel("diff. (%)")
[…]
