### 2023-10-17 b8cafa5
Expose installation validation
