+++ src/triumvirate/__init__.py
@@ -27,6 +27,9 @@ cosmological analyses.
 """
 from importlib.metadata import PackageNotFoundError, version
 
+# Installation validation
+from ._valid_install import validate_installation  # noqa: F401
+
 
 __copyright__ = 'Copyright 2023, Mike S Wang & Naonori S Sugiyama'
 __date__ = '2023-10-04'
