+++ src/triumvirate/_twopt.pyx
@@ -12,13 +12,13 @@ from libcpp.string cimport string
 import numpy as np
 cimport numpy as np
 
-from triumvirate._particles cimport CppParticleCatalogue, _ParticleCatalogue
-from triumvirate.dataobjs cimport (
+from ._particles cimport CppParticleCatalogue, _ParticleCatalogue
+from .dataobjs cimport (
     Binning, CppBinning,
     LineOfSight,
     PowspecMeasurements, TwoPCFMeasurements, TwoPCFWindowMeasurements
 )
-from triumvirate.parameters cimport CppParameterSet, ParameterSet
+from .parameters cimport CppParameterSet, ParameterSet
 
 
 cdef extern from "include/twopt.hpp":
