### 2023-10-18 addb821
Add array checks
