+++ src/triumvirate/_arrayops.py
@@ -1,0 +1,402 @@
+"""
+Array Operations (:mod:`~triumvirate._arrayops`)
+==========================================================================
+
+Perform array operations.
+
+.. autosummary::
+    ShapeError
+
+1-d arrays:
+
+.. autosummary::
+    extrap_lin
+    extrap_loglin
+    extrap_pad
+
+2-d arrays:
+
+.. autosummary:
+    extrap2d_lin
+    extrap2d_loglin
+    extrap2d_pad
+
+"""
+from collections.abc import Sequence
+
+import numpy as np
+
+
+class ShapeError(ValueError):
+    """Exception raised when the array shape is incorrect.
+
+    """
+    pass
+
+
+# ========================================================================
+# Extrapolation
+# ========================================================================
+
+def _check_1d_array(a,
+                    check_sorted=False, check_lin=False, check_loglin=False):
+    """Check the input array is a 1-d array, possibly satisfying
+    additional conditions.
+
+    Parameters
+    ----------
+    a : array of float
+        Input array.
+    check_sorted : bool, optional
+        If `True` (default is `False`), check the input array is sorted.
+    check_lin : bool, optional
+        If `True` (default is `False`), check the input array is
+        linearly spaced.
+    check_loglin : bool, optional
+        If `True` (default is `False`), check the input array is
+        log-linearly spaced (and all entries have the same sign).
+
+    Returns
+    -------
+    a : 1-d array of float
+        Squeezed 1-d array.
+
+    Raises
+    ------
+    ShapeError
+        When the input array is not equivalently 1-d.
+    ValueError
+        When the input array fails any of the checks specified.
+
+    """
+    a = np.squeeze(a)
+
+    if a.ndim != 1:
+        raise ShapeError("Input array is not 1-d.")
+
+    if check_sorted:
+        if not np.all(a[:-1] <= a[1:]):
+            raise ValueError("Input array is not an increasing sequence.")
+    if check_lin:
+        spacing = np.mean(a[1:] - a[:-1])
+        if not np.allclose(a[1:] - a[:-1], spacing):
+            raise ValueError("Input array is not linearly spaced.")
+    if check_loglin:
+        same_sgn = np.all(a > 0. if a[0] > 0. else a < 0.)
+        if not same_sgn:
+            raise ValueError("Input array contains mixed-sign entries.")
+        log_spacing = np.mean(np.log(a[1:] / a[:-1]))
+        if not np.allclose(np.log(a[1:] / a[:-1]), log_spacing):
+            raise ValueError("Input array is not log-linearly spaced.")
+
+    return a
+
+
+def _check_2d_array(a,
+                    check_sorted=False, check_lin=False, check_loglin=False):
+    """Check the input array is a sorted 2-d array, possibly satisfying
+    additional conditions.
+
+    Parameters
+    ----------
+    a : array of float
+        Input array.
+    check_sorted : bool, optional
+        If `True` (default is `False`), check the input array is sorted.
+    check_lin : bool, optional
+        If `True` (default is `False`), check the input array is
+        linearly spaced.
+    check_loglin : bool, optional
+        If `True` (default is `False`), check the input array is
+        log-linearly spaced (and all entries have the same sign).
+
[…]
