+++ src/triumvirate/_bihankel.py
@@ -4,29 +4,6 @@ Double Hankel Transforms (:mod:`~triumvirate.bihankel`)
 
 Perform Hankel-related double transforms using the FFTLog algorithm.
 
-Array operations
-----------------
-
-.. autosummary::
-    ShapeError
-
-1-d arrays:
-
-.. autosummary::
-    extrap_lin
-    extrap_loglin
-    extrap_pad
-
-2-d arrays:
-
-.. autosummary:
-    extrap2d_bilin
-    extrap2d_biloglin
-    extrap2d_bipad
-
-Transforms
-----------
-
 .. autosummary::
     transform_bispec_to_3pcf
     transform_3pcf_to_bispec
@@ -36,338 +13,6 @@ import numpy as np
 from scipy import interpolate
 
 import triumvirate._fftlog as _fftlog
-
-
-class ShapeError(ValueError):
-    """Exception raised when the array shape is incorrect.
-
-    """
-    pass
-
-
-# ========================================================================
-# Extrapolation
-# ========================================================================
-
-def _check_1d_array(a):
-    """Check the input array is a sorted 1-d array.
-
-    Parameters
-    ----------
-    a : 1-d array of float
-        Input 1-d array.
-
-    Returns
-    -------
-    a : 1-d array of float
-        Input 1-d array.
-
-    Raises
-    ------
-    ShapeError
-        When the input array is not 1-d.
-
-    """
-    a = np.squeeze(a)
-
-    if a.ndim != 1:
-        raise ShapeError("Input array is not 1-d.")
-    # if not np.all(a[:-1] <= a[1:]):
-    #     raise ValueError("Input array is not an increasing sequence.")
-
-    return a
-
-
-def _check_2d_array(a):
-    """Check the input array is a sorted 2-d array.
-
-    Parameters
-    ----------
-    a : 2-d array of float
-        Input 2-d array.
-
-    Returns
-    -------
-    a : 2-d array of float
-        Input 2-d array.
-
-    Raises
-    ------
-    ShapeError
-        When the input array is not 2-d.
-
-    """
-    a = np.squeeze(a)
-    if a.ndim != 2:
-        raise ShapeError("Input array is not 1-d.")
-    # if (
-    #     not np.all([np.all(a_row[:-1] <= a_row[1:]) for a_row in a])
-    # ) or (
-    #     not np.all([
-    #         np.all(a_col[:-1] <= a_col[1:]) for a_col in a.transpose()
-    #     ])
-    # ):
-    #     raise ValueError("Input array is not an increasing sequence.")
-
-    return a
-
-
-def extrap_lin(a, n_ext):
-    """Extrapolate a 1-d array linearly.
-
-    The linear spacing used is the distance between the two points
-    at either edge separately.
-
-    Parameters
-    ----------
-    a : 1-d array of float
-        Input 1-d array.
-    n_ext : int
-        Number of extra elements on either side.
-
-    Returns
-    -------
-    a_out : :class:`numpy.ndarray`
-        Extrapolated 1-d array.
-
-    """
-    a = _check_1d_array(a)
-
-    a_l = a[0] + (a[1] - a[0]) * np.arange(-n_ext, 0)
-    a_r = a[-1] + (a[-1] - a[-2]) * np.arange(1, n_ext + 1)
-    a_out = np.concatenate([a_l, a, a_r])
-
-    return a_out
-
-
-def extrap_loglin(a, n_ext):
-    """Extrapolate a 1-d array log-linearly.
-
-    Parameters
-    ----------
-    a : 1-d array of float
-        Inout 1-d array.
-    n_ext : int
[…]
