+++ src/triumvirate/src/parameters.cpp
@@ -913,8 +913,8 @@ int ParameterSet::print_to_file(char* out_parameter_filepath) {
 
   std::fclose(ofileptr);
 
-  if (trv::sys::currTask == 0) {
-    trv::sys::logger.info(
+  if (trvs::currTask == 0) {
+    trvs::logger.info(
       "Check used-parameter file for reference: %s.", out_parameter_filepath
     );
   }
