### 2023-10-19 450f3aa
Make `hankl` comparison optional
