### 2023-10-19 e213163
Implement extrapolation options and update tests
