+++ src/triumvirate/src/arrayops.cpp
@@ -55,233 +55,184 @@ const char* ExtrapError::what() const noexcept {return this->err_mesg.c_str();}
 
 namespace array {
 
-int check_1d_array(double* a, int N, bool check_lin, bool check_loglin) {
+int check_1d_array(
+  std::vector<double>& a, bool check_lin, bool check_loglin, bool check_sign
+) {
+  std::size_t N = a.size();
+
   if (check_lin) {
     std::vector<double> diff_a(N - 1);
-    for (int i = 0; i < N - 1; i++) {
+    for (std::size_t i = 0; i < N - 1; i++) {
       diff_a[i] = a[i + 1] - a[i];
     }
-    if (!check_isclose(diff_a, diff_a[0])) {
-      return 1;
-    }
+    if (!check_isclose(diff_a, diff_a[0])) {return 1;}
   }
+
   if (check_loglin) {
     std::vector<double> diff_loga(N - 1);
-    for (int i = 0; i < N - 1; i++) {
+    std::feclearexcept(FE_ALL_EXCEPT);
+    for (std::size_t i = 0; i < N - 1; i++) {
       diff_loga[i] = std::log(a[i + 1] / a[i]);
+      if (std::fetestexcept(FE_DIVBYZERO) || std::fetestexcept(FE_INVALID)) {
+        return 3;
+      }
     }
-    if (!check_isclose(diff_loga, diff_loga[0])) {
-      return 2;
+    if (!check_isclose(diff_loga, diff_loga[0])) {return 2;}
+  }
+
+  if (check_sign) {
+    for (std::size_t i = 0; i < N - 1; i++) {
+      if (a[i] * a[i + 1] <= 0.0) {return 3;}
     }
   }
+
   return 0;
 }
 
-void extrap_loglin(double* a, int N, int N_ext, double* a_ext) {
-  // Check for sign change or zero.
-  double dlna_left = std::log(a[1] / a[0]);
-  if (std::isnan(dlna_left)) {
-    throw trv::sys::ExtrapError(
-      "Sign change or zero detected in log-linear extrapolation "
-      "at the lower end."
-    );
-  }
+void extrap_lin(
+  std::vector<double>& a, int N_ext, std::vector<double>& a_ext
+) {
+  // Prepare arrays.
+  int N = a.size();
+  int N_notch_left = N_ext;
+  int N_notch_right = N_ext + N;
+  int N_extrap = N_ext + N + N_ext;
 
-  double dlna_right = std::log(a[N - 1] / a[N - 2]);
-  if (std::isnan(dlna_left)) {
-    throw trv::sys::ExtrapError(
-      "Sign change or zero detected in log-linear extrapolation "
-      "at the upper end."
-    );
-  }
+  a_ext.resize(N_extrap);
 
   // Extrapolate the lower end.
-  if (a[0] > 0) {
-    for (int i = 0; i < N_ext; i++) {
-      a_ext[i] = a[0] * std::exp((i - N_ext) * dlna_left);
-    }
-  } else {
-    for (int i = 0; i < N_ext; i++) {
-      a_ext[i] = 0.;
+  double da_left = a[1] - a[0];
+  for (int i = 0; i < N_notch_left; i++) {
+    a_ext[i] = a.front() + (i - N_notch_left) * da_left;
+  }
+
+  // Fill in the middle part.
+  for (int i = N_notch_left; i < N_notch_right; i++) {
+    a_ext[i] = a[i - N_notch_left];
+  }
+
+  // Extrapolate the upper end.
+  double da_right = a[N - 1] - a[N - 2];
+  for (int i = N_notch_right; i < N_extrap; i++) {
[…]

[…]
