### 2023-10-19 3a45f12
Refine array checks
