### 2023-10-19 a85aedb
Add extrapolation reality note to docs. Update syntax/fix typos
