+++ src/triumvirate/transforms.py
@@ -145,3 +145,156 @@ class SphericalBesselTransform:
         post_samples *= prefactor * parity
 
         return post_sampts, post_samples
+
+
+class DoubleSphericalBesselTransform:
+    """Double spherical Bessel Transform.
+
+    Parameters
+    ----------
+    degrees : (int, int)
+        Degrees of the double spherical Bessel transform.
+    biases : (int, int)
+        Power-law bias indices.
+    sample_pts : array of float
+        Pre-transform sample points in 1-d for both dimensions.
+        Must be log-linearly spaced.  Must be even in length if
+        extrapolation is used.
+    pivot : float, optional
+        Pivot value (default is 1.).  When `lowring` is `True`, this is
+        adjusted if it is non-zero, or otherwise directly calculated.
+    lowring : bool, optional
+        Low-ringing condition (default is `True`).
+    extrap : int, optional
+        Extrapolation method (default is 0) with the following
+        options:
+
+        - 0: none;
+        - 1: extrapolate by constant padding;
+        - 2: extrapolate linearly;
+        - 3: extrapolate log-linearly.
+
+        Any extrapolation doubles the sample size in effect, and
+        assumes the pre-transform samples to be real.
+
+    Attributes
+    ----------
+    degrees : (int, int)
+        Degrees of the spherical Bessel transform.
+    biases : (int, int)
+        Power-law bias indices.
+    size : int
+        Sample size for either dimension.
+    pivot : float
+        Pivot value used for both dimensions.
+
+    """
+    # Representative transform, either vertical (0, across rows)
+    # or horizontal (1, across columns).
+    _rep = 0
+
+    def __init__(self, degrees, biases, sample_pts, pivot=1.,
+                 lowring=True, extrap=0):
+
+        self.degrees = degrees
+        self.biases = biases
+
+        self._fbsjt = tuple(
+            SphericalBesselTransform(
+                degree_, bias_, sample_pts,
+                pivot=pivot, lowring=lowring, extrap=extrap
+            )
+            for (degree_, bias_) in zip(degrees, biases)
+        )
+        self._lowring = lowring
+        self._extrap = extrap
+
+        self._logres = self._fbsjt[self._rep]._logres
+        self._pre_sampts = np.asarray(self._fbsjt[self._rep]._pre_sampts)
+        self._post_sampts = np.asarray(self._fbsjt[self._rep]._post_sampts)
+
+    @property
+    def size(self):
+        """Sample size.
+
+        """
+        return self._fbsjt[self._rep]._nsamp
+
+    @property
+    def pivot(self):
+        """Pivot value.
+
+        """
+        return self._fbsjt[self._rep]._pivot
+
+    def transform(self, pre_samples):
+        """Transform samples at initialised sample points.
+
+        Parameters
+        ----------
+        pre_samples : array_like
+            Pre-transform samples in 2-d.  Assumed to be real
+            if extrapolation is used.
+
+        Returns
+        -------
+        post_sampts, post_samples : array_like
+            Post-transform sample points and samples in 2-d, with
+            `post_sampts` in :func:`numpy.meshgrid` 'ij'-indexing format.
+
+        """
+        inter_samples = []
+        for pre_samples_row in pre_samples:
+            _, inter_samples_row = self._fbsjt[1].transform(pre_samples_row)
+            inter_samples.append(inter_samples_row)
+
+        post_samples_T = []
+        for inter_samples_col in np.asarray(inter_samples).T:
+            _, post_samples_col = self._fbsjt[0].transform(inter_samples_col)
+            post_samples_T.append(post_samples_col)
+
+        post_samples = np.asarray(post_samples_T).T
+
+        post_sampts = np.meshgrid(
[…]
