+++ examples/FFTLog/sj2d_demo.py
@@ -1,4 +1,5 @@
-"""Demonstrate the use of the FFTLog algorithm for Hankel-like transforms.
+"""Demonstrate the use of the FFTLog algorithm for double spherical
+Bessel transforms.
 
 .. attention::
 
@@ -13,23 +14,13 @@ import os.path as osp
 import sys
 import warnings
 from functools import wraps
-from typing import Callable, List, Literal, Self, Tuple, Union
+from typing import Callable, Literal, Self, Tuple, Union
 
 import matplotlib.pyplot as plt
 import numpy as np
-from scipy.interpolate import InterpolatedUnivariateSpline
+from scipy.interpolate import RectBivariateSpline
 
-from triumvirate._fftlog import HankelTransform
-from triumvirate.transforms import SphericalBesselTransform
-
-try:
-    import hankl
-except ImportError:
-    hankl = None
-    warnings.warn(
-        "Could not import `hankl` package. "
-        "Some of the test cases will not be available."
-    )
+from triumvirate.transforms import DoubleSphericalBesselTransform
 
 
 def parse_parameters() -> argparse.Namespace:
@@ -42,22 +33,18 @@ def parse_parameters() -> argparse.Namespace:
 
     """
     parser = argparse.ArgumentParser(
-        description="FFTLog Hankel transform demo."
+        description="Double spherical Bessel transform demo."
     )
 
     parser.add_argument(
         '-t', "--test-case", type=str,
-        choices=['hankel-sym', 'hankel-asym', 'sj-sym', 'sj-cosmo'],
+        choices=['sym', 'asym', 'cosmo'],
         required=True
     )
     parser.add_argument(
-        '--order', type=int, default=0,
-        help="Order of the Hankel transform (default is %(default)d)."
-    )
-    parser.add_argument(
-        '--degree', type=int, default=0,
-        help="Degree of the spherical Bessel transform "
-             "(default is %(default)d)."
+        '--degrees', type=int, nargs=2, default=(0, 0),
+        help="Degrees of the double spherical Bessel transform "
+             "(default is %(default)s)."
     )
     parser.add_argument(
         '--extrap', type=int, nargs='?', const=3, default=0,
@@ -82,10 +69,10 @@ def parse_parameters() -> argparse.Namespace:
     pars = parser.parse_args()
 
     # HACK: Override the degree to be 0 to match the external samples.
-    if pars.test_case == 'sj-cosmo' and pars.degree != 0:
-        pars.degree = 0
+    if pars.test_case == 'cosmo' and pars.degrees != (0, 0):
+        pars.degrees = (0, 0)
         warnings.warn(
-            "Spherical Bessel transform degree overriden to 0 "
+            "Double spherical Bessel transform degree overriden to (0, 0) "
             "when external cosmological multipole samples are referenced "
             "as they are computed for the monopole only."
         )
@@ -113,7 +100,7 @@ def get_sampts_lgrange(centre: float, shorten: float) -> Tuple[float, float]:
 
 
 def get_analy_func_pair(
-    testcase: Literal['hankel-sym', 'hankel-asym', 'sj-sym']
+    testcase: Literal['sym', 'asym']
 ) -> Tuple[Tuple[Callable, str], Tuple[Callable, str]]:
     """Get pre- and post-transform analytical functions for the given
     test case.
[…]
