### 2023-10-23 267c2c4
Refactor extrapolation
