### 2023-10-23 a1f882c
Add analytical test function
