+++ examples/FFTLog/fftlog_demo.py
@@ -18,6 +18,7 @@ from typing import Callable, List, Literal, Self, Tuple, Union
 import matplotlib.pyplot as plt
 import numpy as np
 from scipy.interpolate import InterpolatedUnivariateSpline
+from scipy.special import gamma
 
 from triumvirate._fftlog import HankelTransform
 from triumvirate.transforms import SphericalBesselTransform
@@ -47,7 +48,7 @@ def parse_parameters() -> argparse.Namespace:
 
     parser.add_argument(
         '-t', "--test-case", type=str,
-        choices=['hankel-sym', 'hankel-asym', 'sj-sym', 'sj-cosmo'],
+        choices=['hankel-sym', 'hankel-asym', 'sj-sym', 'sj-asym', 'sj-cosmo'],
         required=True
     )
     parser.add_argument(
@@ -93,7 +94,7 @@ def parse_parameters() -> argparse.Namespace:
     return pars
 
 
-def get_sampts_lgrange(centre: float, shorten: float) -> Tuple[float, float]:
+def get_sampts(centre: float, shorten: float) -> np.ndarray:
     """Get the log-range of sample points.
 
     Parameters
@@ -109,11 +110,15 @@ def get_sampts_lgrange(centre: float, shorten: float) -> Tuple[float, float]:
         Log-range of sample points.
 
     """
-    return tuple((val - centre) / shorten for val in LGRANGE)
+    return np.logspace(
+        *((val - centre) / shorten for val in LGRANGE),
+        round(NSAMP / shorten),
+        base=10
+    )
 
 
 def get_analy_func_pair(
-    testcase: Literal['hankel-sym', 'hankel-asym', 'sj-sym']
+    testcase: Literal['hankel-sym', 'hankel-asym', 'sj-sym', 'sj-asym']
 ) -> Tuple[Tuple[Callable, str], Tuple[Callable, str]]:
     """Get pre- and post-transform analytical functions for the given
     test case.
@@ -128,6 +133,8 @@ def get_analy_func_pair(
         - 'hankel-asym' for asymmetric-form functions for the
           Hankel transform;
         - 'sj-sym' for symmetric-form functions for the spherical
+          Bessel transform;
+        - 'sj-asym' for asymmetric-form functions for the spherical
           Bessel transform.
 
     Returns
@@ -159,6 +166,19 @@ def get_analy_func_pair(
             ), (
                 lambda k, ell=0: (2*np.pi)**(3./2) * k**ell * np.exp(-k*k/2.),
                 r"$g(k) = (2\pi)^{3/2} k^\ell \mathrm{e}^{-k^2 / 2}$"
+            ))
+        case 'sj-asym':
+            return ((
+                lambda r, s=1, **kwargs: r**(s - 3),
+                r"$f(r) = r^{s - 3}$"
+            ), (
+                lambda k, s=1, ell=0, **kwargs:
+                    np.pi**(3./2)
+                    * gamma(s/2. + ell/2.) / gamma(3./2 - s/2. + ell/2.)
+                    * (2./k)**s,
+                r"$g(k) = \pi^{3/2} "
+                r"\frac{\Gamma(s/2 + \ell/2)}{\Gamma(3/2 - s/2 + \ell/2)} "
+                r"\left(\frac{2}{k}\right)^s$"
             ))
         case _:
             raise ValueError(
@@ -257,10 +277,10 @@ class comparison_plot:
             self._ax_comp.set_xlim(*self._xlim)
         if self._ylim:
             self._ax_comp.set_ylim(*self._ylim)
[…]
