### 2023-10-23 6cd8927
Add `mcfit` comparison
