### 2023-10-25 ffdb21a
Unify extrapolation comparison with `mcfit`
