### 2023-10-25 f281a47
Replace backward with forward transform post-kernel to fix array order
