### 2023-10-25 e033da0
Enforce even parity check for extrapolation and update docs. Remove sort order check in extrapolation. Update syntax
