### 2023-10-26 d2f0952
Change default arg values. Add and refactor tests
