+++ tests/test_transforms.py
@@ -3,98 +3,162 @@
 """
 import numpy as np
 import pytest
+from scipy.special import gamma
 
 from triumvirate._fftlog import HankelTransform
-from triumvirate.transforms import SphericalBesselTransform
+from triumvirate.transforms import (
+    DoubleSphericalBesselTransform,
+    SphericalBesselTransform,
+)
 
 
-def fa_analytic(r, mu=0.):
-    return r ** (mu + 1.) * np.exp(-r*r/2.)
+@pytest.fixture
+def hankel_func_pair(request):
+    """Analytical function pair for testing the Hankel transform.
+
+    """
+    if request.param == 'sym':
+        return (
+            lambda r, mu=0., **kwargs: r ** (mu + 1.) * np.exp(-r*r/2.),
+            lambda k, mu=0., **kwargs: k ** (mu + 1.) * np.exp(-k*k/2.),
+        )
+    if request.param == 'asym':
+        return (
+            lambda r, mu=0., **kwargs: r * (1. + r**2)**(-3./2),
+            lambda k, mu=0., **kwargs: k * np.exp(-k),
+        )
+    raise ValueError("Unknown analytical function pair.")
 
 
-def ga_analytic(k, mu=0.):
-    return k ** (mu + 1.) * np.exp(-k*k/2.)
+@pytest.fixture
+def sj_func_pair(request):
+    """Analytical function pair for testing the spherical
+    Bessel transform.
 
-
-def fb_analytic(r):
-    return r * (1 + r**2)**(-3/2)
-
-
-def gb_analytic(k):
-    return k * np.exp(-k)
-
-
-def fsph_analytic(r, ell=0):
-    return r ** ell * np.exp(-r*r/2.)
-
-
-def gsph_analytic(k, ell=0):
-    return (2*np.pi)**(3./2) * k ** ell * np.exp(-k*k/2.)
+    """
+    if request.param == 'sym':
+        return (
+            lambda r, ell=0, **kwargs:
+                r ** ell * np.exp(-r*r/2.),
+            lambda k, ell=0, **kwargs:
+                k ** ell * np.exp(-k*k/2.) * (2*np.pi)**(3./2),
+        )
+    if request.param == 'asym':
+        return (
+            lambda r, s=1, **kwargs: r**(s - 3),
+            lambda k, s=1, ell=0, **kwargs:
+                np.pi**(3./2)
+                * gamma(s/2. + ell/2.) / gamma(3./2 - s/2. + ell/2.)
+                * (2./k)**s,
+        )
 
 
 @pytest.mark.parametrize(
-    "test_case, order, bias, nsamp, lg_range, extrap",
+    "order, bias, nsamp, lgrange_samp, extrap, hankel_func_pair, lgrange_test",
     [
-        ('a', 2, 0, 2**10, [-5., 5.], 0),
-        ('b', 0, 0, 2**11, [-5., 5.], 3),
-    ]
+        (2, 0, 2**10, [-5., 5.], 0, 'sym', [-3., .5]),
+        (0, 0, 2**11, [-5., 5.], 3, 'asym', [-3., .5]),
+    ],
+    indirect=['hankel_func_pair',]
 )
-def test_hankeltransform(test_case, order, bias, nsamp, lg_range, extrap):
+def test_hankeltransform(order, bias, nsamp, lgrange_samp, extrap,
+                         hankel_func_pair, lgrange_test):
 
-    if test_case not in ['a', 'b']:
-        raise ValueError("Invalid test case.")
+    f, g = hankel_func_pair
 
-    pre_sampts = np.logspace(*lg_range, nsamp, base=10, endpoint=False)
+    pre_sampts = np.logspace(*lgrange_samp, nsamp, base=10, endpoint=False)
     transformer = HankelTransform(
[…]
