+++ src/triumvirate/_bihankel.py
@@ -1,288 +1,0 @@
-"""
-Double Hankel Transforms (:mod:`~triumvirate.bihankel`)
-==========================================================================
-
-Perform Hankel-related double transforms using the FFTLog algorithm.
-
-.. autosummary::
-    transform_bispec_to_3pcf
-    transform_3pcf_to_bispec
-
-"""
-import numpy as np
-from scipy import interpolate
-
-import triumvirate._fftlog as _fftlog
-
-
-# ========================================================================
-# Transforms
-# ========================================================================
-
-def transform_bispec_to_3pcf(ell1, ell2, k_in, bk_in, r_out,
-                             n_fftlog, extrap=None, n_extrap=None):
-    r"""Transform bispectrum samples to three-point correlation function
-    (3PCF) samples using double 1-d FFTLog operations.
-
-    Parameters
-    ----------
-    ell1, ell2 : int
-        Multipole degree of the bispectrum.
-    k_in : 1-d array of float
-        Input wavenumbers.
-    bk_in : 2-d array of float
-        Input bispectrum.
-    r_out : 1-d array of float
-        Output separations.
-    n_fftlog : int
-        FFTLog sample number.
-    extrap : {'lin', 'loglin', 'zero'}, optional
-        If not `None` (default), set one of the following options---
-
-        - 'lin': input bispectrum is extrapolated linearly;
-        - 'loglin': input bispectrum is extrapolated log-linearly;
-        - 'zero': input bispectrum is zero padded.
-
-        In each case, `n_extrap` must be set, and the input wavenumbers
-        `k_in` are extrapolated log-linearly.
-    n_extrap : int, optional
-        Extrapolation sample number (one-sided) (default is `None`).
-
-    Returns
-    -------
-    zeta_dict : dict
-        Output 3PCF including its attributes---
-
-        - 'r', 'zeta': post-transform separation and 3PCF samples;
-        - 'r_fftlog', 'zeta_fftlog': FFTLog separation and 3PCF samples;
-        - 'r1_mesh', 'r2_mesh': post-transform separation mesh grid;
-        - 'n_fftlog': FFTLog sample number;
-        - 'ell1', 'ell2': multipole degrees.
-
-    Raises
-    ------
-    ValueError
-        When `n_extrap` is not set but `extrap` is.
-
-
-    .. attention::
-
-        This function assumes the wavenumber samples :math:`k_{1,2}` in
-        the bispectrum :math:`B_{\ell_1 \ell_2 L}(k_1, k_2)` are the same.
-
-    """
-    # Prepare samples.
-    if extrap is not None:
-        if n_extrap is None:
-            raise ValueError("`n_extrap` must be provided if `extrap` is set.")
-        k_sample = extrap_loglin(k_in, n_extrap)
-
-        if extrap == 'lin':
-            bk_sample = extrap2d_lin(bk_in, n_extrap)
-        elif extrap == 'loglin':
-            bk_sample = extrap2d_loglin(bk_in, n_extrap)
-        elif extrap == 'zero':
-            bk_sample = extrap2d_bipad(bk_in, n_extrap)
-        else:
-            raise ValueError(f"Unknown `extrap` option: {extrap}.")
-    else:
-        k_sample, bk_sample = k_in, bk_in
-
-    n_sample = len(k_sample)
-
-    # Perform FFTLog transform.
[…]
