+++ docs/source/apidoc_py.rst
@@ -12,3 +12,5 @@ Python Package
     apidoc_py/triumvirate.fieldmesh
     apidoc_py/triumvirate.threept
     apidoc_py/triumvirate.twopt
+    apidoc_py/triumvirate.transforms
+    apidoc_py/triumvirate.window
