### 2023-10-26 8d64f08
Fix example blocks in docstring
