+++ src/triumvirate/winconv.py
@@ -1,8 +1,16 @@
 """
-Window Convolution (:mod:`~triumvirate.window`)
+Window Convolution (:mod:`~triumvirate.winconv`)
 ==========================================================================
 
+.. versionadded:: 0.4.0
+
+
 Perform window convolution of three-point statistics.
+
+.. autosummary::
+    WinConvTerm
+    WinConvFormulae
+    ThreePointWinConv
 
 """
 from typing import NamedTuple
@@ -15,64 +23,71 @@ from triumvirate.transforms import (
 )
 
 
-# Default convolution formula from Sugiyama et al. (2018) [arXiv:1803.02132].
-_FORMULAE_SUGIYAMA19 = {
-    '000': [
-        ('000', '000', 1),
-        ('000', 'ic', -1),
-        ('110', '110', 1./3),
-    ],
-    '110': [
-        ('000', '110', 1),
-        ('110', '000', 1),
-        ('110', 'ic', -1),
-    ],
-    '202': [
-        ('000', '202', 1),
-        ('202', '000', 1),
-        ('202', 'ic', -1),
-        ('110', '112', 1./3),
-        ('112', '110', 1./3),
-    ],
-    '112': [
-        ('000', '112', 1),
-        ('112', '000', 1),
-        ('112', 'ic', -1),
-        ('022', '110', 2./5),
-        ('202', '110', 2./5),
-        ('110', '022', 2./5),
-        ('110', '202', 2./5),
-    ],
+NAMED_FORMULAE = {
+    # Sugiyama et al. (2018) [arXiv:1803.02132]
+    'sugiyama+18': {
+        '000': [
+            ('000', '000', 1.),
+            ('000', 'ic', -1.),
+            ('110', '110', 1./3),
+        ],
+        '110': [
+            ('000', '110', 1.),
+            ('110', '000', 1.),
+            ('110', 'ic', -1.),
+        ],
+        '202': [
+            ('000', '202', 1.),
+            ('202', '000', 1.),
+            ('202', 'ic', -1.),
+            ('110', '112', 1./3),
+            ('112', '110', 1./3),
+        ],
+        '112': [
+            ('000', '112', 1.),
+            ('112', '000', 1.),
+            ('112', 'ic', -1.),
+            ('022', '110', 2./5),
+            ('202', '110', 2./5),
+            ('110', '022', 2./5),
+            ('110', '202', 2./5),
+        ],
+    },
 }
 
 
-class _WConvTerm(NamedTuple):
+class WinConvTerm(NamedTuple):
     r"""Window convolution term as a tuple of three factors.
-
-    For example, ``term = _WConvTerm('000', 'ic', -1)``, with
-    ``term.deg_Q = '000'``, ``term.deg_zeta = 'ic'``
-    and ``term.coeff = -1``, corresponds to the term
-    :math:`- 1 \cdot Q_{000} \zeta_{\mathrm{ic}}`.
 
     Attributes
     ----------
-    deg_Q : str
-        Window function multipole.
-    deg_zeta : str
-        Unwindowed three-point correlation funciton multipole.
+    ind_Q : str
+        Window function multipole index/indices.
+    ind_Z : str
+        Unwindowed three-point correlation funciton multipole
+        index/indices.
     coeff : float
-        Numerical factor for the convolution term.
+        Numerical coefficient for the convolution term.
+
+    Examples
+    --------
+    The following `term` corresponds to the term
+    :math:`- 1 \cdot Q_{000} \zeta_{\mathrm{ic}}`:
+
+    >>> term = WinConvTerm('000', 'ic', -1)
+    >>> print(term.coeff, term.ind_Q, term.ind_Z)
[…]
