+++ src/triumvirate/transforms.py
@@ -2,6 +2,9 @@
 Hankel-Like Transforms (:mod:`~triumvirate.transforms`)
 ==========================================================================
 
+.. versionadded:: 0.4.0
+
+
 Perform Hankel-like transforms in 1- and 2-d.
 
 .. autosummary::
