### 2023-10-30 acabacc
Refactor `WinConvTerm` and `WinConvFormulae`. Update docstrings
