+++ src/triumvirate/winconv.py
@@ -5,67 +5,112 @@ Window Convolution (:mod:`~triumvirate.winconv`)
 .. versionadded:: 0.4.0
 
 
-Perform window convolution of three-point statistics.
+Perform window convolution of two- and three-point statistics.
 
 .. autosummary::
     WinConvTerm
     WinConvFormulae
+    TwoPointWinConv
     ThreePointWinConv
 
 """
-from typing import NamedTuple
+import warnings
+from collections import namedtuple
+from fractions import Fraction
 
 import numpy as np
+from scipy.interpolate import InterpolatedUnivariateSpline, RectBivariateSpline
 
 from triumvirate.transforms import (
     DoubleSphericalBesselTransform,
+    SphericalBesselTransform,
     resample_lglin,
 )
 
 
 NAMED_FORMULAE = {
+    # Wilson et al. (2016) [arXiv:1511.07799]
+    'wilson+16': {
+        0: [
+            (0, 0, 1),
+            (2, 2, Fraction(1, 5)),
+            (4, 4, Fraction(1, 9)),
+        ],
+        2: [
+            (2, 0, 1),
+            (0, 2, 1),
+            (2, 2, Fraction(2, 7)),
+            (4, 2, Fraction(2, 7)),
+            (2, 4, Fraction(2, 7)),
+            (4, 4, Fraction(100, 693)),
+            (6, 4, Fraction(25, 143))
+        ],
+        4: [
+            (4, 0, 1),
+            (2, 2, Fraction(18, 35)),
+            (4, 2, Fraction(20, 77)),
+            (6, 2, Fraction(45, 143)),
+            (0, 4, 1),
+            (2, 4, Fraction(20, 77)),
+            (4, 4, Fraction(162, 1001)),
+            (6, 4, Fraction(20, 143)),
+            (8, 4, Fraction(490, 2431)),
+        ]
+    },
     # Sugiyama et al. (2018) [arXiv:1803.02132]
     'sugiyama+18': {
         '000': [
-            ('000', '000', 1.),
-            ('000', 'ic', -1.),
-            ('110', '110', 1./3),
+            ('000', '000', 1),
+            ('000', 'ic', -1),
+            ('110', '110', Fraction(1, 3)),
         ],
         '110': [
-            ('000', '110', 1.),
-            ('110', '000', 1.),
-            ('110', 'ic', -1.),
+            ('000', '110', 1),
+            ('110', '000', 1),
+            ('110', 'ic', -1),
         ],
         '202': [
-            ('000', '202', 1.),
-            ('202', '000', 1.),
-            ('202', 'ic', -1.),
-            ('110', '112', 1./3),
-            ('112', '110', 1./3),
+            ('000', '202', 1),
+            ('202', '000', 1),
+            ('202', 'ic', -1),
+            ('110', '112', Fraction(1, 3)),
+            ('112', '110', Fraction(1, 3)),
         ],
         '112': [
-            ('000', '112', 1.),
-            ('112', '000', 1.),
-            ('112', 'ic', -1.),
-            ('022', '110', 2./5),
-            ('202', '110', 2./5),
-            ('110', '022', 2./5),
-            ('110', '202', 2./5),
+            ('000', '112', 1),
+            ('112', '000', 1),
+            ('112', 'ic', -1),
+            ('022', '110', Fraction(2, 5)),
+            ('202', '110', Fraction(2, 5)),
+            ('110', '022', Fraction(2, 5)),
+            ('110', '202', Fraction(2, 5)),
         ],
     },
 }
+"""Named formulae for window convolution.
+
+This is a `dict` where each key corresponds to a named formula below:
+
[…]
