### 2023-10-31 d219048
Refactor array-related exceptions
