### 2023-10-31 1da40f1
Add window convolution classes
