+++ src/triumvirate/winconv.py
@@ -10,8 +10,12 @@ Perform window convolution of two- and three-point statistics.
 .. autosummary::
     WinConvTerm
     WinConvFormulae
-    TwoPointWinConv
-    ThreePointWinConv
+    TwoPointWinConvBase
+    TwoPCFWinConv
+    PowspecWinConv
+    ThreePointWinConvBase
+    ThreePCFWinConv
+    BispecWinConv
 
 """
 import warnings
@@ -21,6 +25,7 @@ from fractions import Fraction
 import numpy as np
 from scipy.interpolate import InterpolatedUnivariateSpline, RectBivariateSpline
 
+from triumvirate._arrayops import SpacingError, _check_1d_array
 from triumvirate.transforms import (
     DoubleSphericalBesselTransform,
     SphericalBesselTransform,
@@ -316,30 +321,27 @@ class WinConvFormulae:
         return ' '.join(terms_str)
 
 
-class TwoPointWinConv:
-    pass
-
-
-class ThreePointWinConv:
-    """Window convolution of three-point statistics.
+class TwoPointWinConvBase:
+    """Generic window convolution of two-point statistics.
 
     Parameters
     ----------
     formulae : str, |formulae_dicttype| or |formulae_type|
         Window convolution formulae.  If a string, it is assumed to be
-        a named formula (e.g. 'sugiyama+18').
+        a named formula (e.g. 'wilson+16').
     window_sampts : 1-d array of float
         Window function separation sample points.
-    window_multipoles : dict of {str: 2-d array of float}
+    window_multipoles : dict of {str: 1-d array of float}
         Window function multipole samples (each key is a multipole).
 
 
     .. attention::
 
-        All convolution assumes that the 2-d window function,
-        three-point correlation function (3PCF) and the bispectrum
-        multipole samples are square matrices, i.e. they are sampled
-        at the same sample points in both dimensions.
+        If the window convolution formulae use `str` or `int` to represent
+        a multipole, the corresponding `dict` of window function multipole
+        samples must use exactly the same key type.  In other words, the
+        type representation of a multipole must be consistent across
+        all inputs.
 
     """
 
@@ -359,287 +361,684 @@ class ThreePointWinConv:
         self._rQ_in = window_sampts
         self._Q_in = window_multipoles
 
-    def conv_bispec(self):
-        pass
 
-    def conv_3pcf(self):
-        pass
-
-    def conv_3pcf_diag(self):
-        pass
-
-
-def wconv_3pcf_diag(formulae, Q, zeta, r_common, r_Q=None, r_zeta=None):
-    r"""Convolve multipoles of the three-point correlation function (3PCF)
-    and the window function in diagonal form using specified formulae.
+class TwoPCFWinConv(TwoPointWinConvBase):
+    """Window convolution of two-point correlation function (2PCF).
 
     Parameters
     ----------
-    formulae : dict of {str: sequence of (str, str, float)} or 'sugiyama+19'
-        Window convolution formulae encoded as a dictionary where for
-        each convolved-multipole degree (the key), the value is
-        a sequence of tuples each corresponding to a window convolution
-        term including the numerical coefficient.
-
-        For example,
-
-        .. code-block::
-
-            formulae = {
-                '000': [
-                    ('000', '000', 1),
-                    ('000', 'ic', -1),
-                    ('110', '110', 1./3),
-                ],
-            }
-
-        computes only the '000' monopole
-
-        .. math::
-
-            \tilde{\zeta}_{000} =
-                1 \cdot Q_{000} \zeta_{000}
-                + (-1) \cdot Q_{000} \zeta_{\mathrm{ic}}
[…]
