+++ examples/FFTLog/fftlog_demo.py
@@ -12,7 +12,8 @@ import argparse
 import os.path as osp
 import sys
 import warnings
-from functools import wraps
+from functools import partial, wraps
+from timeit import Timer
 from typing import Callable, Literal, Self, Tuple, Union
 
 import matplotlib.pyplot as plt
@@ -87,6 +88,10 @@ def parse_parameters() -> argparse.Namespace:
     parser.add_argument(
         '--show-diff', action='store_true',
         help="Show the difference between the transform results."
+    )
+    parser.add_argument(
+        '--time', action='store_true',
+        help="Time the transform without showing a plotted figure."
     )
 
     pars = parser.parse_args()
@@ -273,6 +278,8 @@ class comparison_plot:
         Labels for the x- and y-axes.
     title
         Title for the plot.
+    timed
+        If `True`, time the transform and skip showing the figure.
 
     """
 
@@ -284,7 +291,8 @@ class comparison_plot:
         ylim: Union[Tuple[float, float], None] = None,
         xlabel: Union[str, None] = None,
         ylabel: Union[str, None] = None,
-        title: Union[str, None] = None
+        title: Union[str, None] = None,
+        timed: bool = False,
     ) -> None:
         # Store figure attributes.
         self._comp_type = comp_type
@@ -294,6 +302,7 @@ class comparison_plot:
         self._xlabel = xlabel
         self._ylabel = ylabel
         self._title = title
+        self._timed = timed
 
     def __enter__(self) -> Self:
         # Set up the figure.
@@ -344,13 +353,18 @@ class comparison_plot:
         self._ax_comp.set_title(self._title)
 
         plt.subplots_adjust(wspace=0, hspace=0)
-        plt.show()
+        plt.show(block=not self._timed)
 
     def __call__(self, runner: Callable) -> Callable:
         @wraps(runner)
         def context_wrapper(*args, **kwargs):
             with self:
-                return runner(*args, canvas=self, **kwargs)
+                direct_runner = partial(runner, *args, canvas=self, **kwargs)
+                if self._timed:
+                    timer = Timer(direct_runner)
+                    print(f"Run time: {timer.timeit(number=1):.3f} s.")
+                else:
+                    return direct_runner()
         return context_wrapper
 
 
@@ -365,6 +379,7 @@ def get_testcase(pars: argparse.Namespace) -> None:
     """
     testcase = pars.test_case
     showdiff = pars.show_diff
+    timed = pars.time
 
     @comparison_plot(
         comp_type='analytic',
@@ -372,6 +387,7 @@ def get_testcase(pars: argparse.Namespace) -> None:
         xlim=(1.e-5, 1.e1),
         ylim=(1.e-6, 1.e0),
         ylabel=r"$g(k)$",
+        timed=timed,
     )
     def run_hankel(canvas: comparison_plot) -> None:
         # Get analytical functions.
@@ -465,6 +481,7 @@ def get_testcase(pars: argparse.Namespace) -> None:
             'sj-asym': (1.e-0, 1.e5),
         }.get(testcase),
         ylabel=r"$k^2 g(k)$",
+        timed=timed,
     )
     def run_sj(canvas: comparison_plot) -> None:
         # Get analytical functions.
@@ -536,6 +553,7 @@ def get_testcase(pars: argparse.Namespace) -> None:
         xlim=(2.5e0, 2.5e2),
         ylim=(-.25e-0, 2.25e0),
         ylabel=r"$r \xi(r)$",
+        timed=timed,
     )
     def run_sj_cosmo(canvas: comparison_plot) -> None:
[…]
