### 2023-11-01 ea8e810
Refactor timer
