### 2023-11-01 aa06dd3
Fix doc example output string
