### 2023-11-01 9076264
Optimise FFTW plan flag. Update syntax
