### 2023-11-01 127dd1e
Refactor FFTW arrays (avoid `memcpy`)
