### 2023-11-01 cc984a0
Fix outer-layer extrap output array trimming
