### 2023-11-01 bd4903d
Refactor FFTW plans
