+++ src/triumvirate/src/field.cpp
@@ -39,9 +39,12 @@ namespace trv {
 // Life cycle
 // -----------------------------------------------------------------------
 
-MeshField::MeshField(trv::ParameterSet& params) {
+MeshField::MeshField(
+  trv::ParameterSet& params, bool plan_ini, const std::string name
+) {
   // Attach the full parameter set to @ref trv::MeshField.
   this->params = params;
+  this->name = name;
 
   trvs::logger.reset_level(params.verbose);
 
@@ -59,7 +62,33 @@ MeshField::MeshField(trv::ParameterSet& params) {
     trvs::update_maxmem();
   }
 
-  this->initialise_density_field();  // likely redundant but safe
+  this->reset_density_field();  // likely redundant but safe
+
+  // Initialise FFTW plans.
+  if (plan_ini) {
+#if defined(TRV_USE_OMP) && defined(TRV_USE_FFTWOMP)
+    fftw_plan_with_nthreads(omp_get_max_threads());
+#endif  // TRV_USE_OMP && TRV_USE_FFTWOMP
+
+    this->transform = fftw_plan_dft_3d(
+      this->params.ngrid[0], this->params.ngrid[1], this->params.ngrid[2],
+      this->field, this->field,
+      FFTW_FORWARD, FFTW_MEASURE
+    );
+    this->inv_transform = fftw_plan_dft_3d(
+      this->params.ngrid[0], this->params.ngrid[1], this->params.ngrid[2],
+      this->field, this->field,
+      FFTW_BACKWARD, FFTW_MEASURE
+    );
+    if (this->params.interlace == "true") {
+      this->transform_s = fftw_plan_dft_3d(
+        this->params.ngrid[0], this->params.ngrid[1], this->params.ngrid[2],
+        this->field_s, this->field_s,
+        FFTW_FORWARD, FFTW_MEASURE
+      );
+    }
+    this->plan_ini = true;
+  }
 
   // Calculate grid sizes in configuration space.
   this->dr[0] = this->params.boxsize[0] / this->params.ngrid[0];
@@ -76,9 +105,76 @@ MeshField::MeshField(trv::ParameterSet& params) {
   this->vol_cell = this->vol / double(this->params.nmesh);
 }
 
-MeshField::~MeshField() {this->finalise_density_field();}
+MeshField::MeshField(
+  trv::ParameterSet& params,
+  fftw_plan& transform, fftw_plan& inv_transform,
+  const std::string name
+) {
+  // Attach the full parameter set to @ref trv::MeshField.
+  this->params = params;
+  this->name = name;
 
-void MeshField::initialise_density_field() {
+  trvs::logger.reset_level(params.verbose);
+
+  // Initialise the field (and its shadow field if interlacing is used)
+  // and increase allocated memory.
+  this->field = fftw_alloc_complex(this->params.nmesh);
+
+  trvs::gbytesMem += trvs::size_in_gb<fftw_complex>(this->params.nmesh);
+  trvs::update_maxmem();
+
+  if (this->params.interlace == "true") {
+    this->field_s = fftw_alloc_complex(this->params.nmesh);
+
+    trvs::gbytesMem += trvs::size_in_gb<fftw_complex>(this->params.nmesh);
+    trvs::update_maxmem();
+  }
+
+  this->reset_density_field();  // likely redundant but safe
+
+  // Initialise FFTW plans.
+  this->transform = transform;
[…]
