### 2023-11-02 d38f3b0
Refactor `FieldStats` FFTs
