+++ src/triumvirate/src/field.cpp
@@ -1446,7 +1446,7 @@ double MeshField::calc_grid_based_powlaw_norm(
 // Life cycle
 // -----------------------------------------------------------------------
 
-FieldStats::FieldStats(trv::ParameterSet& params){
+FieldStats::FieldStats(trv::ParameterSet& params, bool plan_ini){
   this->params = params;
 
   this->reset_stats();
@@ -1464,10 +1464,35 @@ FieldStats::FieldStats(trv::ParameterSet& params){
   // Calculate mesh volume and mesh grid cell volume.
   this->vol = this->params.volume;
   this->vol_cell = this->vol / double(this->params.nmesh);
+
+  // Set up FFTW plans.
+  if (plan_ini) {
+    this->twopt_3d = fftw_alloc_complex(this->params.nmesh);
+
+    trvs::gbytesMem += trvs::size_in_gb<fftw_complex>(this->params.nmesh);
+    trvs::update_maxmem();
+
+#if defined(TRV_USE_OMP) && defined(TRV_USE_FFTWOMP)
+    fftw_plan_with_nthreads(omp_get_max_threads());
+#endif  // TRV_USE_OMP && TRV_USE_FFTWOMP
+    this->inv_transform = fftw_plan_dft_3d(
+      this->params.ngrid[0], this->params.ngrid[1], this->params.ngrid[2],
+      this->twopt_3d, this->twopt_3d,
+      FFTW_BACKWARD, FFTW_MEASURE
+    );
+
+    this->plan_ini = true;
+  }
 }
 
 // An empty destructor is redundant but left here for future implementations.
-FieldStats::~FieldStats() {}
+FieldStats::~FieldStats() {
+  if (this->plan_ini) {
+    fftw_destroy_plan(this->inv_transform);
+    fftw_free(this->twopt_3d); this->twopt_3d = nullptr;
+    trvs::gbytesMem -= trvs::size_in_gb<fftw_complex>(this->params.nmesh);
+  }
+}
 
 void FieldStats::reset_stats() {
   std::fill(this->nmodes.begin(), this->nmodes.end(), 0);
@@ -1960,17 +1985,12 @@ void FieldStats::compute_ylm_wgtd_2pt_stats_in_config(
 
   // Set up 3-d two-point statistics mesh grids (before inverse
   // Fourier transform).
-  fftw_complex* twopt_3d = fftw_alloc_complex(this->params.nmesh);
-
-  trvs::gbytesMem += trvs::size_in_gb<fftw_complex>(this->params.nmesh);
-  trvs::update_maxmem();
-
 #ifdef TRV_USE_OMP
 #pragma omp parallel for
 #endif  // TRV_USE_OMP
   for (int gid = 0; gid < this->params.nmesh; gid++) {
-    twopt_3d[gid][0] = 0.;
-    twopt_3d[gid][1] = 0.;
+    this->twopt_3d[gid][0] = 0.;
+    this->twopt_3d[gid][1] = 0.;
   }  // likely redundant but safe
 
   // Compute shot noise--subtracted mode powers on mesh grids.
@@ -1995,25 +2015,14 @@ void FieldStats::compute_ylm_wgtd_2pt_stats_in_config(
 
         pk_mode -= sn_mode;
 
-        twopt_3d[idx_grid][0] = pk_mode.real() / this->vol;
-        twopt_3d[idx_grid][1] = pk_mode.imag() / this->vol;
+        this->twopt_3d[idx_grid][0] = pk_mode.real() / this->vol;
[…]
