### 2023-11-02 a99e4c9
Add plan reuse to `FieldStats`. Update comments and commented code
