### 2023-11-02 d40f810
Enhance logging and tracking
