+++ src/triumvirate/src/threept.cpp
@@ -107,7 +107,7 @@ double calc_bispec_normalisation_from_particles(
 double calc_bispec_normalisation_from_mesh(
   ParticleCatalogue& particles, trv::ParameterSet& params, double alpha
 ) {
-  MeshField catalogue_mesh(params, false);
+  MeshField catalogue_mesh(params, false, "`catalogue_mesh`");
 
   double norm_factor =
     catalogue_mesh.calc_grid_based_powlaw_norm(particles, 3);
@@ -290,7 +290,7 @@ trv::BispecMeasurements compute_bispec(
 #endif  // TRV_USE_OMP && TRV_USE_FFTWOMP
 
   // Compute common field quantities.
-  MeshField dn_00(params);  // δn_00(k)
+  MeshField dn_00(params, true, "`dn_00`");  // δn_00(k)
   dn_00.compute_ylm_wgtd_field(
     catalogue_data, catalogue_rand, los_data, los_rand, alpha, 0, 0
   );
@@ -300,7 +300,7 @@ trv::BispecMeasurements compute_bispec(
 
   double vol_cell = dn_00.vol_cell;
 
-  MeshField N_00(params);  // N_00(k)
+  MeshField N_00(params, true, "`N_00`");  // N_00(k)
   N_00.compute_ylm_wgtd_quad_field(
     catalogue_data, catalogue_rand, los_data, los_rand, alpha, 0, 0
   );
@@ -366,7 +366,7 @@ trv::BispecMeasurements compute_bispec(
         // ·······························································
 
         // Compute bispectrum components in eqs. (41) & (42) in the Paper.
-        MeshField G_LM(params);  // G_LM
+        MeshField G_LM(params, true, "`G_LM`");  // G_LM
         G_LM.compute_ylm_wgtd_field(
           catalogue_data, catalogue_rand, los_data, los_rand, alpha,
           params.ELL, M_
@@ -375,8 +375,8 @@ trv::BispecMeasurements compute_bispec(
         G_LM.apply_assignment_compensation();
         G_LM.inv_fourier_transform();
 
-        MeshField F_lm_a(params);  // F_lm_a
-        MeshField F_lm_b(params);  // F_lm_b
+        MeshField F_lm_a(params, true, "`F_lm_a`");  // F_lm_a
+        MeshField F_lm_b(params, true, "`F_lm_b`");  // F_lm_b
 
         if (params.form == "diag") {
           for (int idx_dv = 0; idx_dv < dv_dim; idx_dv++) {
@@ -612,14 +612,16 @@ trv::BispecMeasurements compute_bispec(
         // ·······························································
 
         // Compute shot noise components in eqs. (45) & (46) in the Paper.
-        MeshField dn_LM_for_sn(params);  // δn_LM(k) (for shot noise)
+        MeshField dn_LM_for_sn(params, true, "`dn_LM_for_sn`");  // δn_LM(k)
+                                                                 // (for
+                                                                 // shot noise)
         dn_LM_for_sn.compute_ylm_wgtd_field(
           catalogue_data, catalogue_rand, los_data, los_rand, alpha,
           params.ELL, M_
         );
         dn_LM_for_sn.fourier_transform();
 
-        MeshField N_LM(params);  // N_LM(k)
+        MeshField N_LM(params, true, "`N_LM`");  // N_LM(k)
         N_LM.compute_ylm_wgtd_quad_field(
           catalogue_data, catalogue_rand, los_data, los_rand, alpha,
           params.ELL, M_
@@ -869,7 +871,7 @@ trv::ThreePCFMeasurements compute_3pcf(
[…]
