### 2023-11-02 65e96a6
Pre-update release notes
