### 2023-11-03 e5c5457
Remove cdef-members of Cython classes from docstring
