### 2023-11-03 fbcd198
Remove Cython-generated docstring artifacts
