### 2023-11-06 #5571 @MikeSWang
I can confirm the following:

> ## Additional Author Tasks After Review is Complete
> * [x]  Double check authors and affiliations (including ORCIDs)
> * [x]  Make a release of the software with the latest changes from the review and post the version number here. This is the version that will be used in the JOSS paper. ---> [Version 0.3.0](https://github.com/MikeSWang/Triumvirate/releases/tag/v0.3.0)
> * [x]  Archive the release on Zenodo/figshare/etc and post the DOI here. ---> [10.5281/zenodo.10072128](https://zenodo.org/doi/10.5281/zenodo.10072128)
> * [x]  Make sure that the title and author list (including ORCIDs) in the archive match those in the JOSS paper.
> * [x]  Make sure that the license listed for the archive is the same as the software license.
