### 2023-11-06 fcf7274
Add CITATION and update contribution guidelines
