+++ CITATION.md
@@ -1,0 +1,63 @@
+Please cite the following JOSS article (in BibTeX),
+```tex
+@article{Wang:2023a,
+    author = {Wang, Mike Shengbo and Beutler, Florian and Sugiyama, Naonori S.},
+    year = {2023},
+    title = {{Triumvirate: A Python/C++ package for three-point clustering measurements}},
+    journal = {J.~Open~Source~Softw.},
+    volume = {},
+    number = {},
+    pages = {},
+    doi = {},
+    eid = {arXiv:2304.03643},
+    archivePrefix={arXiv},
+    eprint = {2304.03643},
+    primaryClass={astro-ph.IM},
+}
+```
+and the following Zenodo archive (in BibTeX),
+```tex
+@software{Wang:2023b,
+    author = {Wang, Mike Shengbo and Beutler, Florian and Sugiyama, Naonori S.},
+    year = {2023},
+    title = {{Triumvirate: A Python/C++ package for three-point clustering measurements}},
+    publisher = {Zenodo},
+    version = {0.3.0},
+    date = {2023-10-04},
+    doi = {10.5281/zenodo.10072129},
+    url = {https://doi.org/10.5281/zenodo.10072129},
+}
+```
+
+In addition, please consider citing the following MNRAS articles (in BibTeX):
+```tex
+@article{Sugiyama:2019,
+    author = {Sugiyama, Naonori S. and Saito, Shun and Beutler, Florian and Seo, Hee-Jong},
+    year = {2019},
+    title = {{A complete FFT-based decomposition formalism for the redshift-space bispectrum}},
+    journal = {Mon.~Not.~R.~Astron.~Soc.},
+    volume = {484},
+    number = {1},
+    pages = {364--384},
+    doi = {10.1093/mnras/sty3249},
+    eid = {arXiv:1803.02132},
+    archivePrefix={arXiv},
+    eprint = {1803.02132},
+    primaryClass={astro-ph.CO},
+}
+
+@article{Sugiyama:2018,
+    author = {Sugiyama, Naonori S. and Shiraishi, Maresuke and Okumura, Teppei},
+    year = {2018},
+    title = {{Limits on statistical anisotropy from BOSS DR12 galaxies using bipolar spherical harmonics}},
+    journal = {Mon.~Not.~R.~Astron.~Soc.},
+    volume = {473},
+    number = {2},
+    pages = {2737--2752},
+    doi = {10.1093/mnras/stx2333},
+    eid = {arXiv:1704.02868},
+    archivePrefix={arXiv},
+    eprint = {1704.02868},
+    primaryClass={astro-ph.CO},
+}
+```
