### 2023-11-06 #5571 @MikeSWang
Thanks @ivastar for the suggestions! I have made these edits and updated the paper in the repo.
