+++ .devcontainer/notice_welcome.txt
@@ -1,0 +1,23 @@
+👋 Welcome to `Triumvirate` Dev Container in GitHub Codespaces!
+
+🚦 Contribution guidelines:
+
+    https://github.com/MikeSWang/Triumvirate/blob/main/CONTRIBUTING.md
+
+🔧 Codespace container:
+
+    This codespace launches into a Conda base environment managed by
+    Miniconda. It also contains the Conda environment 'trv-dev' set up by
+    `micromamba` with all build, runtime and test dependencies installed
+    from the 'conda-forge' channel. The VS Code editor is pre-configured
+    with extensions for C++/Python/Cython development and documentation.
+
+To build a developmental version of the project code, run the following
+in the terminal:
+
+    micromamba activate trv-dev
+    make clean
+    make -j -O install useomp=true
+
+This will produce the binary executable at build/bin/triumvirate and
+install an editable version of the Python package.
