+++ .devcontainer/Dockerfile
@@ -1,0 +1,3 @@
+FROM mcr.microsoft.com/devcontainers/miniconda:3
+
+COPY .devcontainer/conda-env.yml /tmp/conda-tmp/conda-env.yml
