### 2023-11-06 f326dc2
Refactor container creation
