+++ .devcontainer/onCreate.sh
@@ -1,0 +1,1 @@
+sudo cp .devcontainer/notice_welcome.txt /usr/local/etc/vscode-dev-containers/first-run-notice.txt
