+++ CONTRIBUTING.md
@@ -42,7 +42,7 @@ if you are adding new features.
 
 ## Development Container
 
-[![Codespaces](https://github.com/codespaces/badge.svg)](https://github.com/codespaces/new?hide_repo_select=true&ref=main&repo=355723559)
+[![Codespaces](https://github.com/codespaces/badge.svg)](https://codespaces.new/MikeSWang/Triumvirate?hide_repo_select=true&ref=main)
 
 A developement container (dev container) is made available through
 GitHub Codespaces. It provides a web-based cloud development environment
